[roll] Roll fuchsia [dhcp] Test NS3 client + NS2 server
Amends the integration tests to be able to choose the netstack versions
for the DHCP server and client independently, and adds expectations
for the newly-run combinations.

Also introduces an `annotate` utility to `netstack_testing_common`
to make it easy to add debug prints to a long-running future to aid in
determining where a test is getting stuck.

Multiply: dhcp-integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849838
Original-Revision: 832e4e40c6d18e9e490513356afc36b3c297f16e
GitOrigin-RevId: 2e559366765fd0ca44b9fea7fb4062d5b670a14a
Change-Id: Icefb41eec08cc2dd65dc9845ba4565ef3c3ff2b2