[roll] Roll fuchsia [components][routing] Employ "route bundles"
Routing operations for use, offer, and expose now act on a RouteBundle
type, which embodies a group of one or more routing declarations, either
a singleton (a single route with a unique source) or an aggregate
(multiple routes or one route with a collection source).

This is useful because in some cases it is possible to aggregate multple
capabilities together, and we want the router to have a standard way to
represent this.

Currently, aggregation is implemented by "lookahead" logic that doesn't
mesh well with the of the algorithm. This change makes the handling of
aggregates a regular part of route traversal and reduces the number of
places that need to handle aggregation individually.

This change does not introduce any new forms of aggregation, but without
it, introducing more types of aggregation will be harder. We plan to
base these types of aggregation upon this change:

- Aggregate multiple collections
- Aggregate collections and static children together
- Map an aggregated expose to a single exposed directory entry

Original-Bug: 4776
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845021
Original-Revision: b0443d223eceec9463f49bd47a50dbaf0776664f
GitOrigin-RevId: d76e2a1bbe38a488d76e9dd027adc740c1a925b9
Change-Id: Ie4681ae0bcd2c3738abb46f7a4b9f57e827f2e62