[roll] Roll fuchsia [driver-runner] Sometimes place boot drivers in pkg drivers collection
The placement of drivers based on the package it originates from is too
simple and leads to dependency cycles where a driver in boot-drivers may
end up depending on pkg-drivers. By taking into consideration the
collections of the parents the driver binds to into consideration when
choosing a collection we can ensure that we avoid such dependency loops.

Additionally, the kHost option from the Collection enum was removed as
it was unused.

Original-Bug: 125674
Test: fx test driver-runner-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850436
Original-Revision: 1d2d8a78d5151c3726c7100b4b27da641cb55551
GitOrigin-RevId: 0f92ecbc0609c92cf9cd70a5df71916606d06cdc
Change-Id: I829011f5cfaa2f413d268eb2211d2e753d40c63a