[roll] Roll fuchsia [zither] Forward both enum member value and expression
This change fixes a bug in which the expression in FIDL used to define
an enum member value was forwarded to the backend as a raw numeric
value: in the case of an OR'ed expression, this would give a bogus
binding (not being able to account for and transform any the other
elements making up the expression). As we do for constant bindings, here
we forward both the reduced numeric value as well as the original
expression (when it differs meaningfully) to the backends: the value is
used to generate the binding, while the expression is giving as a
trailing comment for documentation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850198
Original-Revision: 7bc879a89bac1324579f70b0b551796c236fe322
GitOrigin-RevId: 3415dd01d10c38a3360b99d400453aaf8ab45f8c
Change-Id: Ie7497192a3a18425e838790cf9d7d71d6b9aa10b