[roll] Roll fuchsia [Honeydew] Use tempdir for tracing.terminate_and_download() unit test
This change is a follow-up to fxr/841098. It makes the following changes:

* The trace file name is now lower case.
* The unit test now uses a temporary directory.
The previous code used the /tmp/ working directory for the unit test. This could lead to problems if the unit test was run multiple times in a row, as the trace files would be overwritten. The new code uses a temporary directory for the unit test. This ensures that the trace files are not overwritten, and that the file will be deleted after unit tests execution.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851285
Original-Revision: 5a3c1cd2d56cc4832c5a9803f534c16f4deec48c
GitOrigin-RevId: bf37bf6853f75fd32826f13c26063db6014ca216
Change-Id: I3f774e223faadfd54481e818a7502aa055a0ebae