[roll] Roll fuchsia [cf] Make use of Config* types from decl library.
The fuchsia.component.config library is being removed, this
change begins that migration.

The new types are ABI-compatible with the old ones, so as
long as everything builds this change should be a no-op.

The RealmBuilder protocol is not updated because of difficulty
with OOT versioning of the Dart client library, and this change
adds conversion functions to cm_rust to allow the Rust client
library for RB to interoperate with other libraries.

Original-Bug: 96254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850817
Original-Revision: d6263dadafe2d111199c4faabf768f04a2f8c066
GitOrigin-RevId: 265280e8403f60c9c5b8df59493fb225ff70b2fb
Change-Id: I82d5799ecf5226d79706326b527dfd87e0a615a0