[roll] Roll fuchsia [python] Add excludes to pyrightconfig.json
When I open a python file in fuchsia.git, pyright tries to index the
entire workspace, which for me is 351403 python files. It then pins a
core for the forseeable future and remains generally unresponsive until
it happens to index the file that I'm working in.

By excluding prebuilts, third_party, and build dirs the size of the
workspace goes down to 547 python files, and pyright can finish a
complete workspace check in under 15 seconds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851543
Original-Revision: 3903cacaef3a8a5bdc437cd4878aeaea667a41eb
GitOrigin-RevId: ca424272bd8278e4ab96aba9f1334c9e350ef6ee
Change-Id: I43c180370e518d039993c28d7d8a9afd36666178