[roll] Roll fuchsia [mock-ddk] Integrate with driver dispatchers
This change adds integration between the mock-ddk's MockDevice and the
driver runtime's driver dispatcher. Now the mock device will
automatically have a driver dispatcher available on it.
There will be two modes of operation for this dispatcher.

With driver runtime:
The driver runtime will manage the driver dispatcher, so it can block
safely. This is a new feature that can be enabled by changing
FakeRootParent() to FakeRootParentWithDriverRuntime(). If a user is
using this mode, then they have to post the tasks that touch their
device onto the dispatcher for thread safety.

Without driver runtime (this is the default and what is there today):
This is a basic single threaded mode of operation. The dispatcher is set
as the default for the test thread so tests can directly access devices
without posting tasks. It is unsafe to block threads in this mode
(unless the thread blocking is a manual threads not backed by the
driver dispatcher).

Original-Bug: 124464

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824526
Original-Revision: fea97e2565b7a24cd6a08c843e618c9348dc21f2
GitOrigin-RevId: b61837caeeccb75eddf72b41a29205f00c9eb8e2
Change-Id: I075e78642d49d98bf0e3b45742b93a15bc8aff99