[roll] Roll fuchsia [ffx][ssh] Do not write public SSH key file unless needed
This changes the ssh create if needed function to not write
to the public key file unless the private key is new.

This is important for situations where the public key is read-only,
or on a non-writable file system.

This also correctly checks the file permissions for the private key
to make sure it is 0o0600. This change only happens when running
`ffx doctor --repair-keys`.

Original-Bug: 126684
Original-Bug: 126675
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851997
Original-Revision: 222ead9add51ed3e34bf680b1e7b7bbfa6dc7029
GitOrigin-RevId: 6acde93e0b6ed4b2db8f3489856cbc0526652363
Change-Id: I1d8c69791e60d526fc75f9815964af0e4030af71