[roll] Roll fuchsia [tools][cargo-gnaw] extract package_root and project_root prefix check out of project_root calculation
The project_root method determines what the project root for a given
crate is. For some reason, there was a check to ensure that directory
structures of the project and its vendored crates matched what Fuchsia
wanted in the same method. We need to separate these concerns and place
that check elsewhere. It doesn't belong in the project_root method
itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851326
Original-Revision: 9fe717dd1fc33b3dc57df5f690374d00af45c291
GitOrigin-RevId: 2b1662849c124fb4a2acf5d962b06b0adaa07cf9
Change-Id: If15cd641a865753a98a16eab1f4b61027791bf82