[roll] Roll fuchsia [media/audioconsumer] add audio_consumer component
This CL adds an implementation of fuchsia.media.AudioConsumer to
replace mediaplayer in non-smart-display builds and tests. This
implementation supports LPCM streams only (not compressed streams)
for now. This will be adequate for SBC streams, which are decoded
before they are submitted to the consumer. AAC streams, which are
delivered directly to the consumer, will be rejected for now.
Current end-to-end tests use SBC.

Test: new unit tests for all implemented features

Future CLs will include this component in BT end-to-end tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851221
Original-Revision: db37b1f014116b1d17eaa8bab8be35f89d5f5260
GitOrigin-RevId: f277115d551a9adc3d205cb96d4833f5e87c9ce3
Change-Id: Ibf9873c9ce637bd440d160e61e545f8c14851504