[roll] Roll fuchsia [zither][c] Generate README.md
We update the C backend to generate a README.md comprised of
libray-level FIDL documentation. We only need this for the C backend at
this point, but this support can trivially be added to other backends
if/when we care to.

Original-Bug: 111453
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851998
Original-Revision: 6f3f0a8e7b950ae2bda43169e171d87eebe35614
GitOrigin-RevId: d4d96015b3bbbdee9e2162b432d6c0bfec7ed9a0
Change-Id: I805371cdc95717b7cc52d142003376bbe4aefe96