[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Fix hanging test on NVPTX due to lack of warp sync
Previously this wasn't implemented because it's effectively a no-op.
However, this should be safe to emit on sm_60 architectures. It's
important because it carries semantic importance for whether or not
something can be moved. So we should always emit this instrinsic.

Differential Revision: https://reviews.llvm.org/D149923

GitOrigin-RevId: 6b2216aabce6d769eae54f2e1efff93ad2f5dafa
Original-Revision: 2f6056949920ff3667294b14ebcb0074a2b45529
Roller-URL: https://ci.chromium.org/b/8781963278137184993
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852700
Original-Revision: 24f0098a3cb7180a778fabcf15b7e778304b3f79
Change-Id: I1e0bbcea3a6cfe18cc2a2dbc2eb4f8644578d095