[roll] Roll fuchsia [lockdep] Add support for guarding aliased locks.
Add additional lockdep::Guard constructors that accept two lock
capabilities that are aliases of the same underlying lock. The
underlying lock is acquired/validated/released only once, however both
static analysis capabilities are acquired/released to safely and
correctly assert the permitted actions on annotated entities.

This is useful in situations where multiple instances of an object share
a common lock, typically through a common shared resource. An example
scenario is added to the tests to illustrate the efficacy of this
approach and the danger of using lock held assertions for the same
purpose.

Also fix a minor mistake in another test while were here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852776
Original-Revision: 1d398288256d974c2561ef254fefe33ecd90bec0
GitOrigin-RevId: 8e428a68b30cf418699228f42aec44fba6e81efc
Change-Id: I52093fee4b27bee672d640c74fe45a96b2b0155a