[roll] Roll fuchsia [assembly] Only create the archive for the legacy AIB if the product requests it
Only create the legacy AIB for a product when that product requests
it, as opposed to doing so for any release build.  This should save
seconds on smaller builds, and up to a minute on larger builders.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852301
Original-Revision: 53b0ddd236f5c3acee1aa48f54e6663c32b2ba0d
GitOrigin-RevId: 060a72ecac7fb188609c60988083e6e4bfd248b9
Change-Id: I84b97c1aa464c9f7a99f714dffc3ea05b887eac4