[roll] Roll fuchsia [kernel] Fix vmar alignment test flakes
The AlignmentVmar* test cases flake because they try to allocate a VMAR
of size 8GiB with alignment 8GiB, which sometimes is not satisfiable
depending on existing allocations (which are non-deterministic due to
ASLR). While we can't completely eliminate the possibility of this
happening, we can reduce the probability by decreasing the alignment
requirement to 512MiB, as this is enough to satisfy the needs of the test
case but allows for many more available slots and doesn't significantly
increase the runtime of the test.

Original-Bug: 126070
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852698
Original-Revision: 3a918099866d12bc72eee1d725e56f3f5a465a4e
GitOrigin-RevId: 21e98eb8a1e1d9d092141b67017e2bed5f2116bf
Change-Id: If3f0a5074f1d10b2f0dd027f4e2bb3cc5ad1fc68