[roll] Roll fuchsia [kernel][vm] Allocate slot before adding page
CloneCowPageLocked migrates or forks the target page down the clone
chain. It is possible for allocation of a slot to fail in the page list
of a child in the chain, which would cause AddPageLocked to fail.
Pre-allocate the slot so that this cannot happen. This avoids having to
roll back in the case where we are migrating the page and are halfway
through, i.e. we've removed the page from the parent but haven't
inserted it in a child yet.

Original-Bug: 126368
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852258
Original-Revision: edaa46fdabf092c618b61bf2f0b1e7a4b7ee59eb
GitOrigin-RevId: a0a34b83eda299bc03a218747564dcb809d8af9b
Change-Id: I9b62d755571fdb8b1f190540e4ba754d500942e3