[roll] Roll fuchsia [build] Move loops over toolchain-static data out of template
The scopes created in these loops in test_spec() are over data that
does not change for a given toolchain, and so should not be re-
calculated for every instantiation of the template (thousands of
times), but only once when the template is defined by the importing
of the gni file within the context of a given toolchain.

This results in no change to tests.json, and removes 55 seconds of
cpu-time from GN gen, reducing the average time per invocation of
this template from 18ms to 2.5ms.

Run-All-Tests: True

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853116
Original-Revision: 4f34c434b28f2dc36d23fa20b59e2fb8fe98f112
GitOrigin-RevId: ebc662a1eeffaf052f379a9ba21a806cf0e74ea4
Change-Id: I560f421e11d552f3a6d45a754e041c7ae1b0b1c4