[roll] Roll fuchsia [bt][uart] Linearize serial read callbacks
The serial driver will immediately call the callback if there is still
data available to read, so the call becomes recursive.

If the banjo internals add callback frames, a large queue of data
to read can cause this to overflow the stack.

Queue the read on the dispatcher loop instead.

Test: updated test machinery and added test
Test: fx test bt-transport-uart-tests
Original-Bug: b/280626660
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853198
Original-Revision: bef9174821e8b898c38dd59293a222291833fa15
GitOrigin-RevId: 5416ae68d00d64c0b106751c24d3d33c226a73d3
Change-Id: Ic4af578412bb7924d2599b28a8f82af523b6ad06