[roll] Roll scudo [scudo] Fix missing one block in range marking
When a range contains only one block, we may not mark the pages touched
by the block as can-be-released. This happens in the last group and if
it only contains single block.

Also enhance the existing tests and add a new test for testing the last
block.

Differential Revision: https://reviews.llvm.org/D149866

GitOrigin-RevId: c987a98c508d6e651b3fcf70e533b666c79d749e
Original-Revision: f6a0f54c6c6556bbb4a44e5b0874601746062b88
Change-Id: I275d9ae5a8cebbb45af1ed5c1eef1bf2927d7e3b