[roll] Roll fuchsia [starnix] Return cpu mask size from sched_getaffinity()
sched_getaffinity() syscall is supposed to return size of the cpu
bitmask used inside the kernel, but it was returning 0.
Updated it to return valid result and to calculate bitmask based on
the number of CPUs.

Also updated it to validate PID argument.

Original-Bug: 125672
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/852836
Original-Revision: 23834ace637f51334082e501b94d21f96c900f3b
GitOrigin-RevId: 21028bf84fc1ab7a38214091875e664b363d3f4b
Change-Id: I8f3863bbe4c0ae69af8ce0f5ff8d6bc2b5e02b57