[roll] Roll llvm-project/libc [libc] Make the opcode parameter a compile time constant
Currently the opcode is only valid if it is the same between all of the
ports. This is possible to violate if the opcode is places into a memory
location and then read in a non-uniform manner by the warp / wavefront.
Moving this to a compile time constant makes it impossible to break this
invariant.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D150115

GitOrigin-RevId: 8938a351d7ac9b2e137786427ab35550859924c2
Original-Revision: b9f048f4b3253232b6456b6d12baf2215fec2f84
Change-Id: I655ff76cc9a8d68cf96bc17f9caeaf7adc034548