[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Make the opcode parameter a compile time constant
Currently the opcode is only valid if it is the same between all of the
ports. This is possible to violate if the opcode is places into a memory
location and then read in a non-uniform manner by the warp / wavefront.
Moving this to a compile time constant makes it impossible to break this
invariant.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D150115

GitOrigin-RevId: 03716696b66761c4bf7ea3767982541ddd17f35b
Original-Revision: b9f048f4b3253232b6456b6d12baf2215fec2f84
Roller-URL: https://ci.chromium.org/b/8781637741818649937
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853721
Original-Revision: 7fe470a2c57400f6904c8b8c2cff4a1f444042f5
Change-Id: Ia9ad91a5cf6421a194a51a170e565bf948b8bcd5