[roll] Roll fuchsia [forensics][crasher] Fix RISC-V compilation
Use valid assembly syntax for a known invalid instruction.  The
previous code used ARM syntax not available in RISC-V, and the
value it tried to encode is a newer extension instruction, not
one in the range guaranteed to be an invalid instruction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853356
Original-Revision: eff1adb4c952e8ce294188cdce7e83dbbcad036d
GitOrigin-RevId: 65f8d9c1e84a2227e56aabce7aa73456a8ba8c4e
Change-Id: If43d432bd26283eb7c79b2373ca0d367e64211e6