[roll] Roll fuchsia [devmgr_config] Rename devmgr_config to additional_boot_args
The old name has a lot of legacy associated with it which no longer
applies. In truth, this config could likely be phased out entirely, but
that is a larger change. It is easier to simply rename it to be more
accurate for now.

Original-Bug: 91451
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842837
Original-Revision: 154bc1a99185a77ce6d7742e398a1b76a1ee7483
GitOrigin-RevId: 2ea82effc9df1f8749f5e1d2e661719f9e18696c
Change-Id: I767832a18581a5afe81646da60e95523e25d2ebe