[roll] Roll fuchsia [product] wrap lookup and download in get; disable old get
The prior `ffx product get` is renamed to  `ffx product download`.
The `get` command is now essentially removed.

Adds `ffx product get` which does the function of both lookup and
download. This is behind an --experimental flag which bypasses
the message about the tool being renamed.

The experimental version is not for production use and is a proof
of concept example.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841896
Original-Revision: 2ff871cb5836310b158a8902e8bc5c06e8ae1dfc
GitOrigin-RevId: fee3f16eee7b5a955ccc61d2d4aa720d01105b0a
Change-Id: Ib437948f038b24e6a6e30fcfc1536ae5a219d436