[roll] Roll fuchsia [kernel][vm] Remove unused all_vmos_cursors_
The all_vmos_cursors_ held any cursors that were iterating over the
all_vmos_ list and could drop the global VMO lock. The only usage of
this used to be for harvesting accessed bits, which was removed in
fxr/533288 that moved all the harvesting logic over to the MMU side.

Even though the only remaining use of the VmoCursor now is for the
discardable list, leave the VmoCursor class templated to allow for more
cursor types in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854544
Original-Revision: a3eaeb983f76d673b2142e2814c70b5290f98b32
GitOrigin-RevId: e6d446610d260b2991eef6b77bc315e012c2daeb
Change-Id: I8ffcfc82885f4b83431b793bb0fde22149af0cee