[roll] Roll fuchsia [build][rbe] Handle special cases with '.'
When build_subdir == '.', there is no relative path to match,
so skip all pre/post-flight checks.
This makes it easier to manually drive remote actions
outside of an `fx build` for testing.

Handle cases when build_subdir contains a '.', which should
be interpreted as a literal '.', not regex for "any-char".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854736
Original-Revision: 5a205a27ec242e99ae02785d2f43053c67d01549
GitOrigin-RevId: cb51b8222e5f2478474a87f6f722f446d392f8a1
Change-Id: Ib6310733bf1e4af823d9bd63ef21ee5aa110ec8d