[roll] Roll fuchsia [pci] Rename composite device
The `pci-<BDF>-fidl` name for composite devices has been renamed to be
to have the same name as it's primary fragment, the BDF, eg 01:00.0.

This reduces boilerplate in the topological path. Ideally the composite
device wouldn't appear in the topologial path, but that is not possible
yet.

Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744044
Original-Revision: 5afdc0cf19aff98207864834930e611abb843d8b
GitOrigin-RevId: 6cda6ef6eb4310abb27f6e7fe63a8950fc71c939
Change-Id: I751e411be6c1380324b4cd1bda73334ebd930a85