[roll] Roll fuchsia [audio] Remove incorrect driver API documentation comment
Per the driver SignalProcessing API's GetTopology API method:
If GetElements returns one or more elements, GetTopologies must return
one or more topologies.
Hence it was incorrect to say that it is ok to not expose a topology
if topologies are static. We don't repeat the requirement in the
documentation since it is already in the FIDL file.

Original-Bug: 64877

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853121
Original-Revision: 62a21ae4252d2b3c71f7ce107ce2ba48b6bd02f1
GitOrigin-RevId: 8d3b157f3c10a66ea29fd443a5ff70fdb47a6e8c
Change-Id: I48f8d8c240268b35b7f57d2a469be4b682345d36