[roll] Roll fuchsia [bt][a2dp] Restart streams when unpaused
After successfully suspending a stream or not starting the stream
because a permit was not available initially (in contrast to it being
revoked), add a reservation to the queue so that when permit is
available, we will start the stream.

Some minor cleanup of tests as well.

Original-Bug: b/277376263
Original-Bug: 81265
Test: updated tests; fx test bt-a2dp-lib-tests
Test: connect while call in session, finish call and play music
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845636
Original-Revision: 14730c4a471b2bc6b75e2ca32ec64377acea0d46
GitOrigin-RevId: ecb559cd3c1716ff7dbc04ab366034ac07acf533
Change-Id: Iaa83bc313e8525ebbb65ff448dbb6ca7d2ba695a