[roll] Roll fuchsia [perfetto-bridge] Destroy ConsumerAdapter before test fixture stops
Lsan on the hwasan FYI builder caught a leak in the
perfetto_unittests_package test. It looks like the test fixture would
close and the program would end (and thus invoke leak checking) before
the actual dtor for the allocation is called in a Post task. So there's
a brief period of time where the allocation is not owner by a
unique_ptr. If the test fixture ends and closes the loop before the Post
resolves, then the pointer is left dangling.

This ensures the test fixture explicitly destroys the allocation before
it closes.

Original-Fixed: 126702

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854020
Original-Revision: 3407834d19a95d2d2cdc48a551984008d6a4dde4
GitOrigin-RevId: d850ca63e88009af9d66c9b97525c384747ad3f0
Change-Id: I7c0388dded67b813fdc1c2b0bb9f75a1e73eb285