[roll] Roll fuchsia [ffx][config] Basics of fuchsia_env file parsing
Adds a new library that will hold the fundamental logic of fuchsia_env
files written in toml or json5 and the mechanisms to find and read them.

This is just a starting point so is definitely incomplete. More tests
and logic to come.

Original-Bug: 122608
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854481
Original-Revision: 5057affc1f0ddd0730ceb5d3cd4a76cc157acf0d
GitOrigin-RevId: 36fe183313fddd15445a35b502628739a5df8777
Change-Id: I93f3c426aebc08b2fb4f2c357a733c82df391b7c