[roll] Roll fuchsia [devfs] Clone connections instead of using references
The original code used references to the remote for the
device_controller and device_protocol nodes. This means its possible to
use-after-free if the original remote is removed.

Fix this by cloning the remote.
Fix `clone_remote` while we're here to not return a result, since it
cannot fail.

TEST: fx test device_controller_fidl
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853638
Original-Revision: 9e4d806bef90767dd6e5d000173ceeb7443c0f67
GitOrigin-RevId: 0fd305b5ce9f84b3b20efa12bb60378962174efa
Change-Id: I3427d2709ca32987b9fd6cf1bbe4c8ea0fabcbbd