[roll] Roll fuchsia [kernel][uart] Fix TX irq mask/unmask ordering
There is a potential race in some of the uart drivers where a log
writer thread is waiting on a TX interrupt, but the irq thread masks
the interrupt. This can happen when -
1. Log writer thread is waiting on TX interrupt
2. TX interrupt occurs and the interrupt handler signals blocked threads
3. Before the interrupt handler could set the TX mask, the log writer
   thread loops back, unmask's TX interrupt and block again on the
   event.
4. The interrupt handler now masks the TX interrupt
5. TX interrupt will no longer be triggered and hence the logger thread
   will be blocked forever

This CL fixes the issue by first masking the interrupt in the irq
handler and then signaling the blocked thread.

Original-Bug: 126785
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853877
Original-Revision: 688263cc9914cb6017dfe09506289b4ebc4180b8
GitOrigin-RevId: bc707294808740f93447c68e1159e87aa19309de
Change-Id: I0c481e292a59eba0ecdfc1b871b67bf4a3af3e45