[roll] Roll fuchsia [virtualization] Implement PL011 LCR
Zircon has started to write to this register which is causing test
failures since the vmm is exiting in response to the unimplemented
register.

This is a trivial implementation that simply makes LCR read/write
without implementing any specific behaviors. This is sufficient to get
the PL011 to work with Zircon guests.

Test: fx test virtualization-core-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855248
Original-Revision: 4db8621b32d00acff717ddd112d7f011c1e6df0d
GitOrigin-RevId: 32d44505c8faf609fe14e29d1f6300cb780f1816
Change-Id: I4918216a2e11273371321708a512bf3790b54711