[roll] Roll fuchsia [starnix] Use interruptible locks for append
Before this CL, a write would take a RwLock to ensure that O_APPEND
write were atomics. It means that if a append write was blocked, and a second
write was started, it could not be interrupted before the first one
has finish, even if it should have been interrupte because of signals.

This CL introduce locks that can be interrupted, and uses it for the
append lock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855296
Original-Revision: 81c553e7898fe9f476df7b7e9ac4efc4901f1537
GitOrigin-RevId: 826893ef261ca68d5a76dda95e93ac77be22fd38
Change-Id: I0929d8cd40bf471ad9c397b45a3375f404a1c266