[roll] Roll fuchsia [scrutiny][verify][routes] Correctly handle `--response-level`
`ffx scrutiny verify routes` supports a `--response-level` option, but
it has been handled inconsistently. The option would be passed to the
controller, causing errors, errors and warning, errors and warnings and
capability summaries, or errors and warnings and full route descriptions
to be encoded in the controller's response.

However, the verifier that consumes the controller's response was
ignoring all non-error data, yielding no change in the overall tool's
behaviour when `--resoponse-level` was set.

This change reports a JSON snippet and human-readable messages for all
data returned by the controller. Errors or warnings that are not
filtered out by allowlists result in a non-zero status code. Other
information is printed, but, in the absence of errors or warnings,
yields a status code of 0.

Original-Bug: 103258
Original-Bug: 120171

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851776
Original-Revision: 379efdc1ed97c622cc72e449387744619678183e
GitOrigin-RevId: 01c1ee1a7d166a1f4afcf1f8b6768621875245af
Change-Id: I613823b5f2f672eac7705956188c8537f1ad77ea