[roll] Roll fuchsia [netstack3] Set socket device if addrs require
If the source or destination address of an IP socket have link-local
scope (but are not loopback), set the address of the device in the
socket definition to the egress device. Use this device in TCP and UDP
code to set the bound device for sockets.

Original-Fixed: 125733
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855858
Original-Revision: a7bca82e53933dc0f90a82d4645d925e697b9ac1
GitOrigin-RevId: 8839cc27f4573c8e2f871b3408d4913adebf80e1
Change-Id: I0b5967aca63df985073a32ca7390f1bc3e0cae1a