[roll] Roll fuchsia [gigaboot] Fastboot flash maps between new and legacy partition names
Some devices use different naming conventions for the expected
partitions.
The 'old-style' convention expects the following partitions:
* zircon-{a,b,r}
* vbmeta-{a,b,r}
* misc
* fvm-fuchsia

While the 'new-style' expects the following:
* zircon_{a,b,r}
* vbmeta_{a,b,r}
* durable_boot
* fvm

This presents a coordination problem among tools, infrastructure, and
individuals: which naming convention do we expect on a particular
device, and how do we encourage the new style while allowing the old style?

During the transition to the new style, fastboot flash adopts a
lenient strategy: accept either the old name or the new name and
dynamically map that to the actual name on the device.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856681
Original-Revision: 418c66615d37ec11af6d0ddfd5dec38ef1a35384
GitOrigin-RevId: 2a751ebed3e0e85a5f24a1d4bae5baf958ea625d
Change-Id: I2352249ecd94a96b85da89a82a62dc804bf48803