[roll] Roll fuchsia [lowpan][openthread_fuchsia] Make receiving TREL packets reentrant.
OpenThread may try to send packets out on the TREL interface directly
from the call that is handling a packet being received on the TREL
interface. This reentrancy was causing the `RefCell` that holds the
TREL implementation to panic. This change makes it so that the `RefCell`
is only borrowed immutably (instead of using `borrow_mut()`) when
processing I/O that might be lead to reentrant calls.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855824
Original-Revision: 2742a12c7f044f4d14e04102c04133138461bed7
GitOrigin-RevId: 4c085470a3f7085255ea6713b22b25a99e84c6a7
Change-Id: I2f29d6d0d27397487dbd95a4980d63ef60156ce3