[roll] Roll fuchsia [kernel][riscv] cast down to a 32bit value when using 32bit percpu writer
In the write patch to the percpu pointer, previously the code was
casting up to a 64bit size, which was causing the compiler to emit zero
extension code sequence every time. Since the sw instruction ignores the
top of the register there's no reason to up cast it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856021
Original-Revision: 5e59f385d30850438e3f04043ced0ad0afb7a284
GitOrigin-RevId: 3cd0b2c98304140f842ae885a451bdbaff482cd0
Change-Id: I0d3c25f7aaa9433eee1275d0d1a6c2ddc1c5228e