[roll] Roll fuchsia [netstack] Support add_subnet_route
Supports the `add_subnet_route` AddressParameter option by adding
the on-link subnet route when the address is added and removing it when
the address is removed.

Note that this approach is not smart at all, and doesn't implement
any reference-counting semantics, so it could unexpectedly clobber
a manually-configured route. Instead, it records whether the subnet
route was _newly_ added by this option; if so, the route is deleted
when the address goes away.

Original-Fixed: 125716
Multiply: interfaces-admin-fidl
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855256
Original-Revision: d2189234df4fb3bddc901ea1915c0080eff23990
GitOrigin-RevId: 22edbe34cc41fbe2cc80325f1077b772c0d1f12b
Change-Id: I465612715225b8ec82274afcf841919e1ad46441