[roll] Roll fuchsia [bazel] Add --publish_only flag to publish_packages
In-tree repository management can sometimes conflict with bazel repo
management, particularly when the remote workflow is used or when the
version of ffx in the SDK is different from the one in the Fuchsia tree.

Add a --publish_only option to publish_packages.py to allow publishing
to an existing amber repo, so the other steps in package publishing can
be handled manually or through fx.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854997
Original-Revision: d2cd7bacd9e36a460a89d130abd7d6487a8495e7
GitOrigin-RevId: 8eca166b4aeb0847d06c5e7df91fed804ae7fcd1
Change-Id: I76931185e93039bd973693349fee072f2c7d1757