[roll] Roll fuchsia [log][validator] Refactor validator into its own package
The goal of this is to have a package that we can use OOT when we move
the Dart libraries OOT. That way we can write the test OOT and use the
validator package built in tree.

This approach relies heavily on subpackages and allows to define a new
test by creating a new puppet and a new package using the test topology.
When moving the dart stuff OOT we won't be able to reuse the test logic
(as we are doing right now), but that should be fine as it's a very tiny
main that leaves all heavy logic to the validator component which we'll
provide to the dart test oot.

I think there's potential for generailization here and integration with
TestSuite. I think this pattern could serve as a based for a generalized
conformance testing framework.

Original-Bug: b/240154194

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855396
Original-Revision: b72cf7d3b196e4494e7e40c7d331aefe5039f206
GitOrigin-RevId: 5042878645eb3466a3584d339e094281407ec2c7
Change-Id: I521a8b837d0fc6e1393738a87f1668bb199562f3