[roll] Roll fuchsia [product] avoid creating empty out dir in pbms lib.
The set up for download created an empty output directory. We've
since moved on to a strategy where the data is downloaded to a
temp location and renamed/moved into position.

The empty directory was mostly harmless, but proved to be an issue
in specific mounted devices. Not creating the empty directory fixes
the issue.

Original-Bug: 127332
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856917
Original-Revision: 1264a73f5d20f7accf401c07bba86c87338ed4bc
GitOrigin-RevId: d790604a017c6e310e49b771314cddd6cf29e61f
Change-Id: I3ebeb9b11c98399a1a013ea9a83636b003416f1b