[roll] Roll fuchsia [build] Fix hermetic_inputs_action.py depfile generation.
Ensure that the depfile's parent directory exists before writing it.

This is necessary to remove the use of stamp files in our GN-generated
Ninja build plans. As of now, the parent directory is created because
of the depfile. This will no longer be the case once [1] is submitted
to GN and `no_stamp_files = true` is set in our `.gn` file.

+ minor cleanups and properly handle depfile quoting.

[1] https://gn-review.googlesource.com/c/gn/+/12866

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857183
Original-Revision: 655348036da80fdc5362cb945528046cf823d372
GitOrigin-RevId: 61d81c24efb68f0d9b69ca05f969c29e3000f30c
Change-Id: I3589ce31708d0acd2a316e07c19bc187503c1518