[roll] Roll fuchsia [time][security] remove timekeeper from scrutiny allowlist
The FIDL protocols named in the allowlist are not essential
for RealmBuilder as used by the timekeeper component.

We remove it for this reason.

However, this requires modifying the `offer` for the unneeded
protocols by making them `from: "void"`. Since they are unused
this should make no matter to the correct operation of
timekeeper, but will remove the need for allowlisting it.

Original-Fixed: 115634
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857317
Original-Revision: 094172fd196fd7c4c359092e2855cf9cb9968911
GitOrigin-RevId: 9c6fdd1d92357457678298bf7e96ae35ecc0cbad
Change-Id: Ifc5ecc05ba6871b9ccd79ede0190d3cb9c0071d6