[roll] Roll fuchsia [dash-launcher][fidl] Rename LaunchWith[Pty|Socket] methods
The protocol will be extended with a method for exploring packages,
this change establishes a naming scheme convenient for the new method
(ExplorePackageOverSocket).

Even though the FIDL is not in the SDK, it is called by ffx and so
the renaming needs to be done with a soft migration. The old names can
be deleted once ffx binaries built with the old names no longer need
to be supported.

Original-Bug: 127374
Original-Bug: 126945
Original-Bug: 127317
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857291
Original-Revision: 059d145fe878d8f32b8d030bc86b0d00198bcf33
GitOrigin-RevId: c98fad3b58af0798e2540d85e9046a3beab800b8
Change-Id: Iefb557ef259b8bc9db4df00c0d45c7e682b3cedc