[roll] Roll fuchsia [bootserver_old] Fix paving broken in old bootserver
In change Id7f9669ac8a38583d14a9462302446fa1e32b45a a regression was
introduced into the old C bootserver, and was not noticed until it was
included in an SDK roll by clients which consume the C bootserver
because infra primarily makes use of the Go bootserver (which did not
regress).

The specific failure is that NETBOOT_FVM_FILENAME was redefined to a
different string (and NETBOOT_FVM_SPARSE_FILENAME was added and most
references updated to use that macro value), but the place that
referenced it in bootserver.c was not updated.  This resulted in a
bootserver that would attempt to send "<<image>>fvm.blk" instead of
"<<image>>sparse.fvm", and the (previous) version of zedboot on the
device did not have any handling for that TFTP filename.  As a result,
the device would fail to pave with the old bootserver, including through
`pave.sh`.

Original-Bug: b/277612348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858038
Original-Revision: 8547367312218199b53cfad0498cd7f8060cbf57
GitOrigin-RevId: bc1780d79b0e07431d354e464f0ffc08f5d3e663
Change-Id: I3ce8dc0eb3552055153330c12b67412126f19ec2