[roll] Roll fuchsia [ffxutil] Increase outer timeout around ffx daemon stop.
Sometimes the daemon fails to stop because it hits the context deadline
first. However, uploading the CAS outputs also then fails probably
because some cleanup is still happening in the background so files that
it tries to map to the outputs end up not existing when it actually
tries to upload. Temporarily extend the outer timeout so that hopefully
the bot won't fail to upload the outputs and we can see the logs to
diagnose what's happening.

Original-Bug: 127370
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857788
Original-Revision: 7c8ec9be3fc3e16fd81b7fb03432d350299c442f
GitOrigin-RevId: fa43b172163762b8e9e21967c653a476ff60623b
Change-Id: I9e3784bfa85c7efbde384fdd5eb5aba623085e95