[roll] Roll fuchsia [reference_doc] Support "rename" attribute
The semantics of this attribute is similar to that of serde.

This is used to rename "stop_timeout_ms" to "__stop_timeout_ms" when
rendered to documentation. It looks like we use "__stop_timeout_ms" in
CML in production instead of "stop_timeout_ms".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857636
Original-Revision: cc993c47a7bfbe013b9d788a3757d4fbf0f3d539
GitOrigin-RevId: 821ddd37eb84cf1b489d822cc3733f4c67add9d0
Change-Id: I7c3520ec75aa97928f7b4e99c2bfe0b472d2ba40