[roll] Roll fuchsia [console-launcher] Remove hub and hub-v2 from shell namespace
The v1 hub will be removed when appmgr is deleted, and the v2 hub
is deprecated in favor of `ffx component` and the on-device
`component` shell tool.

In particular, see the `component explore` subcommand to accessing a
component's namespace and outgoing (/out) directory.

Original-Bug: 107482, 100621
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853842
Original-Revision: 976c611163d909ba3d813757d94145fe27a72285
GitOrigin-RevId: b884dae74e5332243bd9ff7639e5ecb49fe670dc
Change-Id: Id584c2f6e1655cd0d514e4c015198110f4b89d48