[roll] Roll fuchsia [starnix] Add POSIX_WRITABLE for RDONLY
The Fuchsia open flags and the Linux open flags do not line up
perfectly. On Fuchsia, flags are hierarchical, which means you need
RIGHT_WRITABLE on a directory in order to open files within that
directory as writable. In Linux, you always open directories as RDONLY.

To resolve this mismatch, Fuchsia has POSIX_WRITABLE which is ignored
for most files but inflated to RIGHT_WRITABLE for directories. When
translating a RDONLY or RDWR flag to Fuchsia, other clients, such as
FDIO, add in POSIX_WRITABLE. We need to do the same thing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858037
Original-Revision: cf9bb12172512f1f2027a1bc4513d7070684b795
GitOrigin-RevId: 219b78168a80c55db1fda84429924566a0cb7265
Change-Id: I01ea5ac17821f310fff1b44c74e5d630190cfb48