[roll] Roll fuchsia [ffx][emu] Add instance name default to default config
This adds the default value for `emu.name` to the data file
which is used as the default configuration for ffx.

This is done as a form of documentation, so all the values
which can be configured are printed when running `ffx config get`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858101
Original-Revision: 5dbee8768dcdaa89e2afb3ff133c5da7e1626679
GitOrigin-RevId: b70ffcfa06e55d038af5d0297c3f9edb86bfc9b7
Change-Id: Ic3dca41a3aa48d3c477411ade667a82fc8dbfa69