[roll] Roll fuchsia [sshd-host] Remove hub and hub-v2 from SSH namespace
The v1 hub will be removed when appmgr is deleted, and the v2 hub
is deprecated in favor of `ffx component` and the on-device
`component` shell tool.

In particular, see the `component explore` subcommand to accessing a
component's namespace and outgoing (/out) directory.

Original-Bug: 107482, 100621
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856693
Original-Revision: 435912b7f7d2c113992e292c80d26d382c95c81c
GitOrigin-RevId: 42cb33aadd374c3d06b81e8a6280287b01488671
Change-Id: Ic475ca78e3919bf86d12c0ad2dc7bf86c8f5d2ec