[roll] Roll fuchsia [stdcompat] Fix std::erase template deduction error
https://reviews.llvm.org/D149543 changed headers such that
vector now transitively includes string. This caused std::string's
std::erase to become a candidate for overload resolution of
std::erase. This was using the template directly and not relying
on argument deduction to weed the std::erase for string out. This
substitution was an error outside of the immediate context of
std::erase so it was not subject to SFINAE.

Instead, just use the left hand side of the assignment to do normal
template argument deduction to choose the correct std::erase.

Original-Fixed: 127378
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858236
Original-Revision: a83c56a0705e6b291a79922f4d16e0eba04ed92e
GitOrigin-RevId: 3f6693a4e28cc258f7f59fba8fd31e04ec5c3fcf
Change-Id: I5e53a5a69ff92447d32aa465448cbcd878b6462b