[roll] Roll fuchsia [starnix] Avoid code injection on module changes
In restricted executor, since the Linux thread and the starnix kernel
are running in the same Zircon thread, it's possible to issue a software
breakpoint in the starnix kernel directly to notify a module change.
This uses less hack and results in a cleaner code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858017
Original-Revision: 3d0f21f45edcb8e7eec4136034b75eb40d07dece
GitOrigin-RevId: f8a02bbf457e1e83a83804392494698a9b0ce1bd
Change-Id: I7e2c7dda51dc71fc645d92675f8ceb2838f63912