[roll] Roll fuchsia [fonts] Create font asset provider targets
These targets are directly usable in AIBs for software assembly.
Each product will define one, and software assembly will pull
those in to provide the fonts resources outside the `config-data`
package.

This approach will allow us to remove the dependency of `config-data`
to the fonts subsystem, leading to eventual removal of `config-data`
altogether.

This is technically a suboptimal way to produce such a bundle, since
it limits the choice of fonts to fuchsia.git (and in-tree petals) only.
We should be able to transition in the future to a method that is fully
governed from the software assembly `ffx` plugin, but at the moment we
don't need that, and also we can migrate from here to that approach
whenever we like.

Hence, the first step is `config-data` deprecation.

Original-Bug: 122855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858496
Original-Revision: ff4dd4bc19e60f5a3cee10054c054f8c44e8ac1b
GitOrigin-RevId: 790ad9edd2b1e42b248dabd48be39cb10acb06b4
Change-Id: I58cd77657933b4e307453646aa10d803ce4810a6