[roll] Roll fuchsia [debugger] Improve error message when downloads fail
Explicitly print a not found error in the case where a BuildID is not
found on any of the currently configured symbol servers (as opposed to
other types of downloading errors).

Point users to the ZXDB bug template if their symbols are not found or
otherwise not working. With the new information, we should be able to
help debug better.

Original-Fixed: 126624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857422
Original-Revision: 70ec1ad76aa9ac39e97aa713d4ee8cf861768c13
GitOrigin-RevId: 12313fc38b2667be32a56a95f810ea5c7229115d
Change-Id: Ib3e780231a13c18120bc9bea353fc2b93e48e607