[roll] Roll fuchsia [driver_manager] Specify collection name to runner directly
Because package name is not the only consideration when choosing which
collection to place the driver, we shouldn't rely on using the package
type to determine the collection name in the runner code. This was
missed by the last attempt to remedy this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857977
Original-Revision: ed67892747aedbfab40333998d393e3dc95df444
GitOrigin-RevId: c0d3710b478c9bc9ed13f3a17f36b5b6407cd330
Change-Id: I0f7a57e2d0baa388410c0585336253c4e1afbf6e