[roll] Roll fuchsia [kernel][riscv64] Revamp syscall dispatch using a lookup table
The old method of syscall dispatch assumed that each trampoline
sequence could fit into 8 bytes. In fact, the trampoiline
sequences were always larger (12 bytes) and then getting relaxed
to smaller (6 bytes) when the jump targets were within 1MiB
distance. When they're not that close, each sequence is too large
and the alignment causes each one to actually be 16 bytes, so
that the indexing calculation is wrong and the dispatch goes
to the wrong trampoline for the syscall number.

This instead uses a lookup table of addresses, stored as 32-bit
offsets to halve the number of cache lines occupied by the table
compared to storing full addresses (that would get relocated).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858556
Original-Revision: 9fde18d634d40943a3b697c6deb2d18d29293a17
GitOrigin-RevId: cbf9dd5e00feb648d106d698ed9da918c4a2e774
Change-Id: I50bbf7d5f1935ff42b63e3c78737ccbef3cbe5fa