[roll] Roll fuchsia [userspace fastboot] add `flash gpt-meta` support
We're adding support for flashing the gpt-meta pseudo-partition to make
it easy to initialize partition tables.

This triggers the exact same logic as `oem init-partition-tables`, but
aliasing it behind a `flash` command allows us to use existing ffx and
flash manifest features without needing to add one-off support for a
custom oem command.

This is only being implemented for userspace fastboot so far, as the
driving use case is to allow `ffx flash` to completely bootstrap a
device from a USB installer. If we ever need this in Gigaboot fastboot
as well, the trick is going to be making sure both the paver and
Gigaboot come up with the exact same partition tables from this command.

Original-Bug: b/282797139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859819
Original-Revision: ba2af17afc7a4667c49355cb562c06bea58cc1c7
GitOrigin-RevId: 7d0d82f6b86cf5c9c0943b22c4561b45e8e5d2a1
Change-Id: I16d7300ac853c0ef49d158d1a1cde0c7d48272b3