[roll] Roll fuchsia [starnix] Show correct mount root in mountinfo
The 4th column in mountinfo is supposed to be the path from the FS
root to the root of the mount (i.e. it shouldn't depend on whether the
same FS is mounted elsewhere). Starnix was returning an empty string
for non-bind mounts. For bind mounts it was the path to the source
dir in the original mount.

Also removed `Mount.origin_mount` as it's not longer needed.

Original-Bug: 79302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858340
Original-Revision: c98739af2f83a392fbfb06052863f21f586574c7
GitOrigin-RevId: 50aff4956e97d1e875b96f098dfef403d5014eaf
Change-Id: I68919ba06fe3753587bdbec47ad6372bbc4862b2