[roll] Roll infra/recipes [tree_closer] Remove ability to set password
This change removes the ability to set a password for the tree closer.
The closer should now always be run using a service account, and so no
password is sent to the googleplex.com request.

The username remains as this is used not for auth(n|z) but as a display
name for the status message.

Original-Bug: b/267492263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/859797
Original-Revision: 013f54657d736011f32b7d4f8cbd5f2af08df5b4
GitOrigin-RevId: 59ce3c3b958a463468c9de337a72d587ccee325b
Change-Id: I83492552b464a3c37f77450c4bf54c8e5d11055d