[roll] Roll fuchsia [ui][testing] ui_test_realm no longer supports config.use_input
use_input was useful for determining whether to expose input-specific protocols in the
test realm. Given that scene_manager is the only possible scene owner,
and it is also always the input owner when in use, this change removes
use_input under the assumption that use_scene_owner suffices for
expressing the same outcomes. We therefore read use_scene_owner in all
the places where use_input would have mattered.

In reality, this means that tests that used to not expose input
protocols now always will. This should be a reasonable outcome for
scene manager-managed realms, where the input protocols are already provided via the scene_manager_scene_with_input manifest, and the fact that scene_manager sans input is not a real production use case. This should also be a reasonable outcome for scenic-only realms, where the only protocol now being exposed is the pointerinjector protocol.

Original-Fixed: 127469
Test: fx test //src/ui/tests
Test: fx lsc presubmit fxr/858616
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858616
Original-Revision: b973d9e953644eab3de28d2450fdad1917852fa1
GitOrigin-RevId: 225ceeb335336846c980ad946ddd735949500a9f
Change-Id: I7f68a9d8579bfb921d74ff2b8c328011ad579653