[roll] Roll fuchsia [ffx] fix bootstrap flashing order
The bootstrap images should always be flashed first before the
bootloaders, since bootstrap is the lowest-level set of images that the
others may depend on.

For example, bootstrap images might contain the GPT which has to be
flashed first so that the bootloaders will go to the right location on
disk.

Original-Bug: b/282797139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859044
Original-Revision: a8a639baf8682fffddc0aa33b7b584e37b4d1bf8
GitOrigin-RevId: d6096d35d84a204c188993a781000de179a5d126
Change-Id: I1ad6e54701ec6f39efe603dcbf028b248c3ddcfc