[roll] Roll fuchsia [assembly] Enable platform AIB for audio
Configure the media subsystem for software assembly. Currently,
we only enable the 'audio_dev_support' AIB (containing the
'virtual_audio' driver and eventually the 'ffx audio' daemon).
We do this on the 'eng' build-type and 'minimal' feature-set.

Thus it is included in (for example):
    core_with_netstack3, terminal, workstation, smart_display_eng
but not in (for example):
    bringup, smart_display_userdebug

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841645
Original-Revision: b24f68284978f958feaff8b43e62942c5bbdfb83
GitOrigin-RevId: 2cb64b5a4767100f16ed53090b725a66a038759b
Change-Id: Ib9e36973b04f914cda4c39c203c0b110826edc5e