[roll] Roll fuchsia [starnix] Refactor file offset handling
File offset is now completely owned and modified by FileOps. File
implementation specify whether the file is seekable or not (and whether
the offset are meaningful or not for file that accepts seek but have not
notion of file position), and FileObject update the offset depending on
result of read/write when needed.
This allows FileOperation to have a single operation that always pass an
offset. This offset will always be 0 for file that do not have
meaningful offsets.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857759
Original-Revision: ae1c066e245380516fd183be6c91262fc8704a25
GitOrigin-RevId: 76cc792dba17fde5a4e9c5678466ec7ea8d727bf
Change-Id: I85c845cb9f6a5ba199efb44b2d65ac5de3594897