[roll] Roll fuchsia [starnix] Remove exception executor
The restricted executor has now reached feature parity with the
exception executor, and is significantly faster.

Maintaining two different executors is a maintenance burden: right now
the exception executor is not even compiled in CQ/CI.

Given that Zircon considers restricted mode part of the kernel's base
functionality, and is committed to supporting it on new architectures
moving forward, the exception executor is unlikely to be useful in the
future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849157
Original-Revision: 429db297f0f5bce1e9dce86b019ea37eb909ae5e
GitOrigin-RevId: b63ce48a437b3215805ae2e9cb1bbe108765f6a1
Change-Id: Icd0243d7f6224860adc9362e7ee6f1a097a2fdb4