[roll] Roll fuchsia [starnix] Extract __kernel_rt_sigreturn offset in vDSO
Compute the offset of sigreturn in the vDSO and store it alongside the
vmo containing the implementation.
Also store the address at which the VvDSO is mapped on the memory
manager and then use this address together with the offset for signal
trampoline address computation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851717
Original-Revision: 7fed30e6733f3b3c5834edede4a193654d61ef8e
GitOrigin-RevId: 194bf0cb43fb20cfe56759a98be6ce400991da19
Change-Id: I7e28baa273d13c5cc22fe33fcf25031d5fc0b3f4