[roll] Roll infra/recipes [btu] Add "upload_to_cipd" flag
In order for a build to upload to CIPD, it must explicitly set
"upload_to_cipd" to True. This allows try/CI/prod builders to have
necessary differences in behavior when uploading, though the recipe will
do all the CIPD package preparation and stop just short of uploading
when the flag is not enabled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/860361
Original-Revision: c13cf62bc17738319b634686604e03b4957ab62b
GitOrigin-RevId: aa2cb66c31b194dd69d4179b9908740c87461c8c
Change-Id: I4d8f088f989f808275e8468909198a2da6b6e5e6