[roll] Roll fuchsia [component][fuzzing] Only collect status with monitors subscribed
This is a small performance CL that adds an `active` flag to the
`MonitorClients` that tracks if any monitors have been added. If not,
the runner will skip (potentially expensive) calls to `CollectStatus` as
part of `UpdateMonitors`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793800
Original-Revision: 0b9370ae4cd7b5ff421e70e2bc37c102a2c5dfea
GitOrigin-RevId: 9a6a1deebb4c954bd35a6c5dac10db6b5c7b444a
Change-Id: I5ee994b6de2577f4f300088024f0c013068a83e8