[roll] Roll fuchsia [assembly][abi_revison validation] Remove unused allowlist mechanism for flag to downgrade errors to warnings
Switch from using an allowlist of package names that are downgraded
to a warning from an error if they fail validation, to a single
option that applies to all abi_revision validation errors.  This
flag currently defaults to warning on abi_revisions validation
errors, and will soon be transitioned to defaulting to being an
error when the remaining failures are addressed.

Original-Bug: 91001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859887
Original-Revision: 4fc24598621e9a8506f7acf182ade4b839554043
GitOrigin-RevId: 7404098f98776e42e02867cc04ba24e604aee24d
Change-Id: I32b185bfd3824a245fbe93db5857392cf46f24e8