[roll] Roll fuchsia [cml] Set rules for aggregation
'Aggregation' is the act of routing multiple capabilities to the same
target with the same name, creating (in effect) an aggregate capability.

Example:

```
expose: [
    {
        service: "fuchsia.examples.Echo",
        from: [ "#boot-drivers", "#pkg-drivers" ],
    },
],
collections: [
    {
        name: "boot-drivers",
        durability: "transient",
    },
    {
        name: "pkg-drivers",
        durability: "transient",
    },
],
```

Enforce rules for when aggregation may be used in cml, which is (for
now):

- The capability must be a service
- The routing directive is `expose` or `offer`
- All sources must belong to the same `from` grouping
- All sources are collections

Later, we will relax these rules to allow children to be sources, once
runtime support for that type of aggregation is landed.

Original-Fixed: 127502
Original-Bug: 4776
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859882
Original-Revision: 92c6c2e9064cfa3682041ddfdb6a9ade0b1f3150
GitOrigin-RevId: c8d5f2b5f217dc7dfe03b032e8a09e2981e74fd7
Change-Id: I4f38649116dfc26c7cdede6fb0945e90ea4bd86a