[roll] Roll fuchsia [driver] TestSyncDispatcher no default ctor
Removes the default constructor of the TestSynchronizedDispatcher. This
helps us keep usages of this consistent.

Also move the Start/Stop functions to be private since they are no
longer called manually but instead through the ctor/dtor.

This change also improve a lot of the documentation provided in the
header so users are less confused between the two options.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859967
Original-Revision: 9d7174800bcfda3e4221fcd0f8602170aad2557b
GitOrigin-RevId: 2c18a978ba6482ce55704d9e1746fb20f2355aa7
Change-Id: I15f9e9213224bd400e987202018cccba138dba6b