[roll] Roll fuchsia [terminal] Add log encoding validator package
This is necessary as part of the effort to move Dart stuff out of
fuchsia.git. We still need conformance testing coverage for the Dart
implementation of logs and inspect. Until RFC-0208
(distributing packages in the sdk) is implemented we'll be placing
the necessary packages in the terminal image to run our tests OOT.

Original-Bug: b/240154194
Original-Bug: 127138
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860317
Original-Revision: 27f08a41a80d16dbfec03a24b5043fbafc40e59e
GitOrigin-RevId: 58c74b81407426f41f31be30a664435bced4ffeb
Change-Id: Ib4067c6a58b69d2186c44b6dc9f5d28c7b6abba8