[roll] Roll fuchsia [remote-control] Support moniker maps on ConnectCapability
This brings the equivalent of RemoteControl/Connect "selector maps" to
ConnectCapability in the form of moniker maps. For now the moniker map
is just a hashamp. Selector maps supported recursive mppaings, for now
moniker maps don't. If we need that in the future we can add it.

The goal of the map is to provide compatibility with older tools when
the moniker in the topology changes. This is aligned with RFC-0169 which
requires newer products to be compatible with older tools for the
duration of the ffx compatibility window.

Just like the current selector map, the moniker map introduced in this
CL is empty.

When we no longer need to support RemoteControl/Connect we can delete
the selector map.

Original-Bug: 126721

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857176
Original-Revision: 010f1bfb6717861b71400273b53be01e96ced2dc
GitOrigin-RevId: 7cbfd3769167a56882db3d3472d26122c7a4a5cd
Change-Id: Ifebaa1c484a06aa8b07f72d5aabdb923feeabc77