[roll] Roll fuchsia [ffx][log] Report counts of dropped and rolled out logs
Looks like this, appended to the end of the log line, in yellow:
[dropped=30202].

This change also ensures that no rolled=N log warnings are generated
when we first read logs from the buffer. It's expected that logs have
been rolled out from the buffer by the time we read and we don't care
about how many. However knowing how many are rolled out as we are
ingesting logs could be useful.

If this information turns out to be too nosiy in the log, we'll add a
flag to optionally add these annotations instead of always adding
them.

Original-Fixed: 127297
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856595
Original-Revision: 8039d4618c590d0d5118cfb6349a2272fd02b691
GitOrigin-RevId: 3f2e22201ff4bb0ff7f052ab041fac5d255cc1fc
Change-Id: I1ee1d989470203d673f8a95c70be6bfd37f861af