[roll] Roll fuchsia [component_manager] Update CPU stats inspect format
We now store the data in arrays. This leads to a much smaller
representation in both the Inspect VMO and the JSON format.

See go/fuchsia-cm-diagnostics-encoding. But as TLDR:

- ~76% less bytes for the measurements in the Inspect VMO
   in Component manager or 50% if we round to the page size
- ~55% less bytes for CPU stats in the inspect.json file
  in snapshots

In preparation for the addition of handle/port counts the data
moves to be under stats instead of cpu_stats.

Original-Fixed: 126219
Original-Bug: b/280805706

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857181
Original-Revision: 15c51074b4fe5359f5461f29abf4866cf54074f6
GitOrigin-RevId: acd4b606814b9dcd84b2a1b0a4e5e343dec82518
Change-Id: Iec4d9de60679abbd327a558ce5187127707a68a0