[roll] Roll infra/recipes [btu] Only attempt to run tests when testing pool is set
Even if the build produces test manifests, we may not always want to run
tests. This could be handled with a new flag like "run_tests", but this
CL chooses to use "testing_pool" as an indication of whether the recipe
should run tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/860482
Original-Revision: e1cce5940062613b258510884a679440ee0c56e2
GitOrigin-RevId: 0af26758f9176e5de033accc9626f95f709a550c
Change-Id: I2bcea5279f1a86513272cbb2bb4bc30191a6b290