[roll] Roll fuchsia [ffx][isolate] Use normal ffx suite test_data
Changes the e2e_emu library to go through the normal ffx test_suite_data
host test data definition rather than bringing in the SDK.

It also changes the api of `Isolate::new_in_test` a bit to allow for
tests that need the sdk present, and to not require the tests to find
and name the build root directory which is almost always just "." in
test runs. It then derives and configures a reasonable location for the
sdk path in there, using the exported core sdk.

Original-Fixed: 127762

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860776
Original-Revision: 660399efdf3ef769bb1d1a08d7c4e1b35971799f
GitOrigin-RevId: 045feaf03187290dbef6d4fa3914b9c365d7404b
Change-Id: Ie5807e928f95d7daa44c59928509f07af7ce88e5