[roll] Roll fuchsia [ffx][fho] Test find_workspace_tools directly
Using an indirection through the toolsuite for this test resulted in it
picking up tools from the sdk, which makes the test fail. Since this is
a unit test it makes more sense to just call the functions we want to
test directly.

Original-Fixed: 127753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861399
Original-Revision: 713e20d61ae5545f4e6b36ca98205a690330bcdf
GitOrigin-RevId: e08c1b2b3d366af42b45b1ec39f49549747dafe1
Change-Id: I42c36359f6ff428f496ea0115f0bbf12328c82ac