[roll] Roll fuchsia [starnix] Refactor sysfs
Use kobject and ktype to represent sysfs objects:
- A kobject is the fundamental unit of sysfs. It is embedded in each
sysfs object. A kobject has a name, a ktype, pointers to its children,
and a pointer to its parent, which allows it to be organized into
hierarchies.
- ktype is the type of object that embeds a kobject. It controls what
happens to the kobject when being created and destroyed.

Refactor sysfs:
- Use the root kobject to create a dynamic directory.
- All implementations of FsNodeOps in sysfs should hold a reference
to its kobject.

Refactor uevent:
- registry.rs is where the dispatch happens for now. It should be
migrated to SysFs in the upcoming CLs.
- All device information should be acquired via the kobject. Other
env states should be kept in the registry.
- More uevent actions are introduced.

Original-Bug: 119437, 121327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854028
Original-Revision: 8e51ddf9f4fe0422a21aca61e4fc809aa10dafc4
GitOrigin-RevId: 5f540a4a44186cb41594410506d5e3153d959249
Change-Id: I86931154e8fc743bfa8dd3141541c295eafa4c92