[roll] Roll fuchsia [kernel] Leak handles when copyout fails
Stop modifying the gHandleTableArena when copying a handle to a user
fails. This introduces the possibility of leaking handles when a caller
provides an invalid user pointer, so this CL modifies the code to
generate an exception. Several tests were doing this, so I've modified
or deleted them to avoid this. This exception will be removed once we
migrate to HandleTableV3.

Future CLs will further reduce the work done by user_out_handle and will
move handle table manipulation into the syscalls themselves.

Original-Bug: 127121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857777
Original-Revision: 94544fb210191c2836731069c9261718d29fce5a
GitOrigin-RevId: d605935094ddf01d1fdaabc174f9dce1c0873674
Change-Id: I2074539e2f45ee46af1c72a681a1363bd402d96f