[roll] Roll fuchsia [amlogic-display] Reland "Use 16x16 SPLIT_BLOCK formats for AFBC by default."
This relands the change fxrev.dev/838617.
The original change uses TILED_HEADER modifier which may cause display artifacts on 1080x1920 displays on VIM3. This reland removes that modifier.

----
Upstream amlogic-display driver requires SPLIT_BLOCK modifier bit
set for AFBC images with 16x16 block, or 4k AFBC images with 32x8
block. If SPLIT_BLOCK is not specified, for large images (e.g.
width = 1920), there'll be blocks / lines of pixels displayed as
random colors.

Besides for better performance, we also added YUV modifier; this format modifier is not used in upstream amlogic
driver but we have verified it's working for vulkan applications.

A side effect of this change is that it breaks display-test, as
the encoding logic of display-test is not compatible with the
YUV modifier we added to the display-driver.
fxbug.dev/125730 tracks the effort to re-enable the test for
amlogic-display.

Test: vkcube-on-fb on astro, no regression (afbc enabled)
Test: vkcube-on-fb on vim3, no regression (afbc disabled)
Test: vkcube-on-fb on vim3, after patching fxrev.dev/838618,
      afbc can be enabled
Original-Bug: 125730
Original-Bug: 124949
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853421
Original-Revision: 42647920c5acb65025a08017784ca17e5934cf15
GitOrigin-RevId: 33e382496239a677588cf26966ebc93473394f15
Change-Id: Iaee60654ff72d11dc94d5e78419d98723b8b20b1