[roll] Roll fuchsia [banjo][display] Config stamp values must be strictly increasing.
Previously, the display controller `config_stamp` values in
ApplyConfiguration() are only required to be unique but not
necessarily increasing. Drivers need to do extra mapping from
config stamp values to their chronological order.

This change adds an extra constraints for config_stamp field in
ApplyConfiguration(); config_stamp must be strictly increasing
so that it can stand for the chronological order of a display
configuration.

Note that the display coordinator (the only client of the banjo
display protocol) is already following this requirements, so this
API change is compatible with the existing client implementation.

Test: API-change only. No expected behavior change.
Original-Bug: 126254
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858890
Original-Revision: 1517fc09c9eabd0d6127473c1d9031e97d71a65f
GitOrigin-RevId: c50e379214a9a6faf0fba55ce791272700805e25
Change-Id: I8bc3b61b77fe9cc966cfcb196d73471f5775397b