[roll] Roll fuchsia [netstack3] Move state out of BoundSocketMap
Move the storage for socket state out of BoundSocketMap and into the
datagram and TCP socket code that depends on BoundSocketMap. The result
is that the BoundSocketMap is used only as a lookup table for
demultiplexing, not as the source-of-truth for socket state. The calling
code needs to be careful to make sure that the address and sharing state
stored for a socket match what's used in the BoundSocketMap.

This will make it easier to store socket state behind references in the
future.

Original-Bug: 126141

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855957
Original-Revision: 6ab5a4570ece5f158222926abafd95be3924542a
GitOrigin-RevId: 0fa1f502c43f6b51748792e69def58caae3855a5
Change-Id: Ic230358287da1e6635b0b8bd5e8701b6446a0480