[roll] Roll fuchsia [display][core] Use DoublyLinkedList to store LayerNode.
Each Layer stores two Nodes of type LayerNode to represent
the Layer in a display's current_layers_ (layers in the applied
config) and pending_layers_ (layers in the config to be applied).

These two layer lists are of type fbl::SinglyLinkedList, so
elements can be only inserted to / popped from the head of the
lists. In order to rebuild a list of nodes using the other list
while keeping the ordering of nodes, the display core has to use
a temporary list to store the nodes in reversed order:

For example, when applying a config, display::Client needs to
build current_layers_ list based on pending_layers_ list.
- In order to do this, display::Client creates a temporary list.
- It pops all the current_nodes to the front of the temporary list
  so that they are sorted in the reversed order in that list.
- It then pops the frontmost node from temporary list, and pushes
  it to the front of the target list (here it's the current layers
  list). Now the nodes are sorted in the correct order.

This breaks the rep invariant of the layer nodes that a pending_node
must be only in pending_layers_ lists, and a current_node must be
only in current_layers_ lists (in the example above, a current_node
is temporarily placed in the temporary list, which breaks the
condition) and makes the code harder to understand in some cases.

This change replaces all SinglyLinkedLists with DoublyLinkedLists.
LayerNode will be DoublyLinkedListable as well.

This change affects LayerNode's (and thus Layer's) memory consumption.
By changing the Node to DoublyLinkedListable, it increases memory usage
of each Layer (each LayerNode now will need 8 more bytes to hold another
node pointer, thus each Layer will need at least 16 more bytes in memory
storage), but makes it easier for us to create layer node lists.

Test: no intended behavior change. display-core-unittests and
vkcube-on-fb passed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822549
Original-Revision: 70d8557343e3bbbdc5c5fd2ee3c14f0667be585a
GitOrigin-RevId: d1745cfcf9ff7676c13bb82c7cc7dc91cf9a7cb2
Change-Id: I3d60d0b855590e23ab459069b4ceb42a39c70399