[roll] Roll fuchsia [driver][compat] Add logging library to ease migration
Most clients want an easy way to migrate existing DFv1 code to DFv2. By
providing a library which provides the existing APIs with new
implementations we lower the cost of migration. Note that while this
code relies on global, it is no worse than what we do in DFv1 already
and what drivers seem to be doing themselves anyways. As long as the
symbol is not exported into the process namespace, we do not need to
worry about drivers resolving to the same symbol.

We may want to consider switching the main driver library to similarly
install a global singleton for improved ergonomics (while continuing to
retain ability to specify the logger).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858308
Original-Revision: 6ede3c6e8dc53c87c92b0aa40b05e36b98fab774
GitOrigin-RevId: 151e4f726f1984f2bfcde93af3829e95e1c85450
Change-Id: I37df9cd6544d6119902a8e98669e09aea50180d2