[roll] Roll fuchsia [radar][proxy] Split radar-proxy into two binaries for injection
Now that there are two AIBs for radar-proxy, the binary can be split so
that the provider proxying and injection code is not included on user
builds. This also requires splitting the test -- shared utilities have
been moved to a header, and test cases to separate source files.

Original-Bug: 100020
Test: fx test radar-integration-test radar-proxy-test
Test: Radar features work on Nelson eng and user builds
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859876
Original-Revision: 8160fb140838f2c31c148605a4d6aa41673b7061
GitOrigin-RevId: a83346de6d58b014e21c876d2cebe8cff47a3b1e
Change-Id: Ia05a38a85643f18048028aa4b55c2a1f2135af65