[roll] Roll fuchsia [netstack3] Expose testutils with feature
When the "testutils" feature is enabled, expose some core testutil types
so they can be used in other crates. This allows other crates to more
easily test against a fake SyncCtx.

To demonstrate the utility of this approach, make the fuzzing binary an
external consumer of Netstack3 core instead of building it as another
version of the crate. This reduces the space of configurations that
netstack3-core is required to support by removing special handling for
fuzzing.

Original-Bug: 125654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857418
Original-Revision: d3196dd357577b760dfb0bec9caa4eedc16bf677
GitOrigin-RevId: b9f82f57174d8f1b94d7fb5ac7fdd9a92ef78f61
Change-Id: I9f8c1a5e1398f533074a670ae234d556685851d5