[roll] Roll fuchsia [fxfs] Only examine golden files from the current build.
Previously we would list the directory and run the golden tests against
all files found with a .zstd extension, but since files added in
previous builds do not get removed in future builds this means that we
may examine files that are not meant to be supported by the subsequent
builds.

Now we add a json manifest file to track exactly which golden images are
expected for the current build and only examine those.

Fix: 99892
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862219
Original-Revision: 24285886d8c2e3e8d4035f3f670a751186a17ead
GitOrigin-RevId: d82c183bfa7d44c382a90dc9fb59789da442859e
Change-Id: Id6a4c1715e575ff996fe80cd59b2228e953c5e12