[roll] Roll fuchsia [ffx][e2e_emu] Run a package server.
Allows developers to iterate on e2e tests locally when their
test's package deps are in the universe set.

To be meaningful, the test for this functionality needs to
resolve a package which is not in the system image produced
by the main build graph. There is a "unit" test in ffx_e2e_emu
for both the successful case with a package server running and
a failure case without one. This ensures that the package is not
available without a package server and is thus in the "universe"
package set as if a developer had added it to their build with
`fx set ... --with //foo`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855101
Original-Revision: 8b61794f55b75268bee8dabe8b1193881c9d575c
GitOrigin-RevId: bc46adbea80f8d4a5026add9549d34064fba2892
Change-Id: If982e01862e1b59b07167bb5c6d9fcace1e997c1