[roll] Roll fuchsia [display-png] Remove gamma table support.
The display API is not supporting setting gamma table
anymore, and we are going to remove this from carnelian
and all the clients that uses this feature.

This removes all the gamma table related code and tests from
display-png binary tool.

Original-Bug: 127817
Test: display_png --background ff0000ff
(Note that the test is failing on ToT when displaying PNG files,
this is tracked in Bug 127903.)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862476
Original-Revision: d3df5dfe08999efebce4728780d0f82236849ba2
GitOrigin-RevId: 6020e199aa43f0bf3e70420d6fccd2ee21b6cfd9
Change-Id: I2db5b2400e69e5e1bf08d2c2b2eb4927ba21beb5