[roll] Roll fuchsia [starnix] Stub MADV_NOHUGEPAGE
This flag is used by madvise calls that are expected to always succeed.

If a madvise fails between a syscall and an assertion on the ERRNO that is expected from that syscall, the test may flake since the ERRNO
has been overwritten.

Original-Fixed: 127894,127934,127933,127930,127911,127927,127899,127925,127923,127922,127921,127919,127916,127914,127913,127910

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862537
Original-Revision: 2251f4737ce2af180f33bff224a262277adad870
GitOrigin-RevId: c0e521816e174f2db5c4975025b5bb72fc6cfd44
Change-Id: I1f0f23a808c50a85541d86f20df5907e1a7c8acb