[roll] Roll fuchsia [pkg-resolver] Log beginning of ephemeral resolve attempt
Devs use this log to determine if an ongoing resolve is the reason
for the device taking a while to do something.

Resolve state is available in inspect, but checking inspect is not
part of the common dev workflow and it's not common knowledge that
pkg-resolver inspect state should be checked for this information.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862152
Original-Revision: 6112a195097b28baa8d4046a9393c8ca6cf37c57
GitOrigin-RevId: 55f74ee932be6b65d7e311ed6e33803557dcf9f0
Change-Id: Ia42654d06f9c659c36963d55e3a5477fc559f61b