[roll] Roll fuchsia [fxfs] Add fsck check for "zombie" files
A zombie file is one which is present in the graveyard (i.e. will be
cleaned up on mount), but there are still links to the file.

Note that these would already emit a RefCountMismatch error, but this
error should be much clearer and easier to debug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863256
Original-Revision: 6b49a9990827b53019427b0f384243c0e4a972d5
GitOrigin-RevId: f8778cb6ede845c79997be23dae8e247e2ee40fa
Change-Id: Ifc2fef17a2b68b594f21094b90802da39271863b