[roll] Roll fuchsia [fxfs] Fixes format error with full seek blocks.
Previously we would `skip()` forward the writer to the end of the block
but when the block was already full we'd accidentally add an extra
block. It isn't necessary to pad out to the end of the block, so just
stop doing that.

Also the WriteBytes trait has been clarified to always append zeroes
when calling `skip()` and the Writer struct has been updated accordingly
to align the behaviours of all implementations.

Original-Bug: 127676

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863236
Original-Revision: cb7907eee8b236dc2ba5263c49b5913fe1a9b4da
GitOrigin-RevId: 81fd932818f37f44843ff27bca9893443de20214
Change-Id: I354cbd8393c03ccd2ee23aca291de5f0e547cd7d