[roll] Roll fuchsia [fxfs] Expand and rename NextKey trait to support merge options.
Replaces the NextKey trait with the LayerKey trait which provides
options to support different necessary merge strategies as well as the
`next_key()` call.

During `find()` we call `push_iterators` which will stop when no
obviously same or adjacent key is found in the current layer. So if
something is not modified long enough adding it to the mutable layer
will drop the old value. By making merge strategies more explicit now
this behaviour will be specifically enabled for the keys that use it
with the default being the cumbersome FullMerge option.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/853919
Original-Revision: 6b19f6c9ede5e0e4a1baa2979150916aedef59c8
GitOrigin-RevId: 3ab18f74801f87431b5c5e2ce6f7b498b3e9fb60
Change-Id: I72aa9122d4e2b9bbe16d2c1e7c56eb17ff58f412