[roll] Roll fuchsia [fxfs] Unwrap keys on the pager executor
The main executor can make blocking syscalls to read the contents of a
file that reenter through the pager executor. The pager executor needs
the unwrapped keys to decrypt the file contents. If the keys are
unwrapped on the main executor then fxfs can deadlock. Moving the
unwrapping of the keys to the pager executor prevents the deadlock.

Original-Fixed: 127056
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862041
Original-Revision: 7b5e76eaec3b288fd5f790c252067b8f58a7462d
GitOrigin-RevId: 8526abe4250f70bc3186d57f4dd88b948958e4ec
Change-Id: Ifd2dbfa3b9450e9b5eb73430ea7bf6d5d9c09f4e