[roll] Roll fuchsia [starnix] Manage extended processor state
This adds explicit management for extended processor state to Starnix.
Extended processor state is architectural state in addition to the
general purpose register state which is managed through Zircon's
restricted mode feature. This includes floating point and vector
registers and floating point control / status registers. Starnix needs
to manage this state in order to safely suspend and resume Linux
programs at arbitrary points of execution and to reset the initial state
for new processes.

There are several possible approaches to managing this state. The
approach that Zircon takes is that it initializes this state when
necessary and saves/restores it on a context switch but it does not
access this state in its system call or exception handling codepaths.
This means that cost of saving and restoring state is avoided in common
codepaths. Adopting this approach in Starnix is challenging since the
program itself and the libraries it depends on including the Rust and
C/C++ standard libraries would need to be compiled specially to avoid
using this state. We may adopt this approach in the future either
wholesale or in some codepaths to optimize their performance.

A variant of this approach would be to partially restrict Starnix's use
of extended processor state and save/restore only that state. In
practice the Rust toolchain makes heavy use of the first few vector
registers for data transfer operations so perhaps we could save only
these and the control/status registers instead of the full state. This
would still be challening to deploy as we would need to consider all of
the libraries Starnix.

Another approach would be to extend the restricted mode feature to also
include saving and restoring extended processor state. This would allow
us to reuse the logic in Zircon used for context switches and grant
access on some architectures (specifically Intel) to supervisor-mode
only instructions XSAVES / XRSTORS. A limitation of this approach is
that it would be more difficult to iterate on the implementation
compared to a purely userspace solution.

This change implements a full state save and restore mechanism in
userspace in a Rust library internal to Starnix. This library is a bit
tricky to use and relies on the Rust toolchain not inserting
instructions that perturb this state in between inline assembly blocks
and on inline routines being always inlined. These properties are not
promised by the Rust reference:

https://doc.rust-lang.org/reference/attributes/codegen.html#the-inline-attribute
https://doc.rust-lang.org/reference/inline-assembly.html#rules-for-inline-assembly

but they are experimentally true and covered by runtime tests in this
change. If this turns out to be problematic we can move the save and
restore logic into the same assembly routine that we use for
restricted_enter and return. For now this logic is exposed in Rust
functions as they are much easier to test and benchmark.

This change also saves and restores the state on all restricted exits
regardless of type. Prior to this change we have not been saving this
state for syscall exits and usually have correct behavior as most
syscalls from Linux are issued through a function call wrapper which is
defined to clobber all or at least the most commonly used vector
registers. It is also uncommon for programs to change the floating point
control register from the default values or to rely on the status
register. Technically this behavior is wrong as a wrapper is not
required by the Linux ABI and this can also leak state across execve().
We may be able to limit the state we save and restore for some type of
exits in the future, such as on architectures that always go through a
defined wrapper.

This costs roughly 100ns per syscall on our NUCs and roughly 50ns per
syscall on a VIM3.

Original-Bug: 126555

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855457
Original-Revision: 612804c28cc8c0d123a00b83e4df316436e10e97
GitOrigin-RevId: f90ce06ab149925ab27fe6a57b458d3d06506ee7
Change-Id: I5e837c5617a6e4fd3d137f38a360cab1c9084fda