[roll] Roll fuchsia [fxfs] support small extended attributes
As an initial implementation, support extended attributes with values
smaller than 256 bytes.

There is a limit on how big we want to make (or even can make) journal
transactions, so beyond a certain size we want to store xattr values in
extents, which is more complicated and will come as a followup.

Original-Bug: 122123
Test: fx test fxfs-platform-test zxio-rust-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851821
Original-Revision: 000845eed750ed6035c3939e9c055d982b00b4ad
GitOrigin-RevId: 4291ff90eb7d6f223663de0986b1ceb0028fb469
Change-Id: I5244813504d974d153733d60062ee5b67a439d75