[roll] Roll fuchsia [fshost] Remove child matchers for data and blobfs
Remove the data and blobfs PartitionMatchers and instead create an
FvmMatcher that explicitly mounts the blob and data partitions. If the
blob partition does not exist, fshost fails. If the data partition does
not exist, reset_fvm_partition will create it.

This approach makes fshost more resilient as it can now handle states in
which the data partition does not already exist. This might happen, for
example, if a device were to lose power in the middle of
reset_fvm_partition after the old data partition had been destroyed but
before the new data partition was allocated. In the old matcher system,
fshost waits for partitions to match against but cannot detect when an
expected partition does not show up. Now, fshost is more explicit about
which partitions it expects to appear in fvm.

Original-Bug: 125255
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/854177
Original-Revision: 65343732279b88eee396da8e55a70a041931023a
GitOrigin-RevId: 91f3cfc3d8e66b8dc99c8c0cc115d69bc33b0dfe
Change-Id: I408fac16f9281eab42d3a7f4e280a49c185da7ba