[roll] Roll fuchsia [bazel][assembly] Use final_zbi in PB metadata
This is a followup on https://fxrev.dev/863156, which might be breaking
a few downstream builders.

There are builder logs suggesting `files.zbi` missing for unknown
reasons. For now, use final_zbi to match previous behavior for GN
assembled PBs, and it is also correctly set for Bazel assembled PBs.

Basically, this should continue to keep `ffx emu` working, and
potentially fix the problems are seeing in downstream builders. The
downside is that we have to live with a unnecessary copy for longer.

Original-Bug: 127949

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863796
Original-Revision: fef0ce244121b270cdcac2bf318dc9cdba9faaf3
GitOrigin-RevId: 335c99ab8fc3e5d1a230b375362c4ff71b9da8b8
Change-Id: I05066110ec6f40a0fb0636c8baaa3ee1a9efd7b0