[roll] Roll fuchsia [build] Define and use toolchain_variant vars for cpp and rust support
define toolchain_variant.supports_[cpp|rust] vars that are then used to
not perform tasks in templates that are being executed in languages
that do not need those tasks to be done (such as asking clang about the
cpu and the target tuple to use.

This uses these new variables in the fidl() template from fidl.gni,
which also causes it to not import certain templates that were
previously available in any BUILD.gn file that imported fidl.gni, and
so a few BUILD.gn files were given import statements to the gni files
that provide those templates.

Original-Bug: 101132

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862916
Original-Revision: 3357db312bc7390b71e27cdf405f639343e111b0
GitOrigin-RevId: 9bb91f66e476cfa282c540a269cc5c13236974f9
Change-Id: I69ce34b8b6b3972c25b6c0f06f8c2d7c4743f57b