[roll] Roll fuchsia [gigaboot] Drain tx queue and free pkt buffers on mDNS broadcast
The UEFI simple network protocol is non-blocking, which means that it
is difficult to know when a packet buffer is no longer in use and can
be freed. Blocking behavior can be implemented by following every call
to transmit with a loop that checks the transmit queue until it is
empty. After the loop terminates, the packet buffer can be safely
freed.

bug: b/281883111

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861525
Original-Revision: 37014e9dca2ad278d96b4ed718ee89e6fb81088b
GitOrigin-RevId: d37e536d5022f1e8a3bcc9a24fb31e5a5f1f3151
Change-Id: I6df94e11fb699b444e06b42420b8de3e9cb58f58