[roll] Roll fuchsia [sysmem][svchost] Move sysmem-connector to miscsvc
This breaks a cycle between drivers and svchost, allow svchost to
provide a strong dep to the driver framework, ensuring it outlives it.
This means that we can continue to process stack traces for crashes that
occur in the driver framework during shutdown.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863724
Original-Revision: 87e33c9522e2008ef40661af42ae1c550779826d
GitOrigin-RevId: 795e6a3f3b7084176d8c57b217a5d2c8910ae957
Change-Id: I018a339c77da87397e53a4762118a7bf96bb40dd