[roll] Roll fuchsia [pci] Allocate BARs with existing addresses first
On NUC11 with an external wifi chip it was found that since the
bootloader only configured BAR2 and provided it the address of the start
of our allocator it became possible for the allocation of BAR0 to
temporarily share the same address until BAR2 was reallocated. Rather
than temporarily disabling either BAR, this resulted in BAR0 not being
mapped at all, while BAR2 continued to operate as normal. To prevent
this corner case we'll allocate the BARs with preconfigured addresses
first.

This is confirmed to fix the wifi issue, but a test to replicate this
situation is still needed. This will delay the fix so it will be
tracked in fxb/128196

Original-Bug: 128196

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862362
Original-Revision: 89d2c4cbdfc46a57e897b0c7c8adb9569e1079b8
GitOrigin-RevId: 17dae090189730a8d49c3b3e3550b5ef405fd516
Change-Id: If8c421facb1d559c91db7fde5b489b9497bff804