[roll] Roll fuchsia [starnix] Enable in-thread exception handling
This enables the in-thread exception handling path in Starnix by
default.

This has a subtle dependency on commit
612804c28cc8c0d123a00b83e4df316436e10e97 in that the thread that
generated the exception is also used for handling it so we must preserve
all architectural state observable by Linux. This adds a test that we
preserve the contents of the first vector register which is used in
several places on the Starnix side.

The threaded exception handling path and configuration is preserved
for now to make debugging issues with this codepath easier. Once the
in-thread exception flow proves stable we can remove the other path.

Original-Fixed: 117302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864159
Original-Revision: b62c38a70573e0d780801ce7bba8f6b537aaa6b2
GitOrigin-RevId: 0a9d9584334075acddce5ed0f86700afe60e6a4d
Change-Id: I8c9b64bbc073624535ee5e4a7ba7744cb200d21d