[roll] Roll fuchsia [starnix] Add /sys/devices/system/cpu/active
This resolves a flake in tcp_socket_test. tcmalloc would sometimes
look at this file, expecting it to be present. If this happened
between a syscall inside of a test assertion and the point at which
the assertion checked the errno, the errno would be overwritten to
ENOENT (since the file was missing).

Multiply: tcp_socket_test

Original-Fixed: 127944, 128001
Original-Bug: 121327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864136
Original-Revision: 5873baf978c5371278b64bf4ebe23a05ca5a9a92
GitOrigin-RevId: bc5c254a381cb1316aa1da4c8dde8c41e6f4b79a
Change-Id: I90a458aa302692f33ff33bb2bad938159eb65f92