[roll] Roll fuchsia [rust-3p] Import loom
This patch imports the library loom 0.5.6. In order to do this, it also:

* Adds generator 0.7.4
* Adds matchers 0.1.0
* Enables the regex-automata feature "regex-syntax"
* Enables the tracing-subscriber feature "env-filter"
* Adds a stub windows rustversion 1.0.12
* Adds scoped-tls 1.0.1
* Adds a stub windows module 0.48.0

loom, generator, matchers, and scoped-tls were approved in OSRB
http://fxbug.dev/107676.

Since the loom crate is intended to be only used in tests, it's marked
as `testonly`. Furthermore, the `generator` indirect dependency uses
assembly to implement green threads, so to reduce our risk of bugs with
it, it also is being marked as `testonly` and restricted to only be used
by `loom`.

Original-Bug: 107676
Original-Bug: 127528

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718991
Original-Revision: f578e714b39ef6e99ce69d147e8ed578f048c71f
GitOrigin-RevId: a8e4d9fe47fabb1d25e5a6bc3de94977fe012618
Change-Id: If8ac008a672b609f314514002f7dfd0cf21990cb