[roll] Roll fuchsia [libc] Disable function type mismatch check for MmapProtExecTest
This particular test copies code onto an executable page then runs it.
However, when checking for function type mismatches, UBSan will check
for a function signature associated with a function that would be
located 8 bytes before the function. Since we only copy the function
itself and not anything before it, this check will lead to a page fault.
This particular copied function would not have been instrumented anyway and
shouldn't have a signature. I think it should be safe to just disable
the check for this one test.

Original-Bug: 128274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864343
Original-Revision: 6b8fac264fd81771848af421e6670e8eba0e2648
GitOrigin-RevId: ed39c723142674564cb3eb072466bb2596df1bd8
Change-Id: Id1ed2644f8dbe78a806abed5a2daa3b1e539e33b