[roll] Roll fuchsia [starnix] Distinguish inode and node identifier in FsNode
Before this CL, the FsNode was using the inode as the unique identifier
of the node, used by the filesystem for caching.

This is problematic for FS like fuse where node needs to be created
before their inode is known. Instead, FsNode now have an unique
identifier that is independent of their inode, but unless the FsNode
implementation changes the inode, it will by default, be set to the node
identifier.

Original-Bug: 125496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864536
Original-Revision: 51b6680ca4165dfbf541883d871b06ba8b1ce97f
GitOrigin-RevId: 149617a43750f6ab76be2a16179f7bdb5ea9e3fb
Change-Id: Ie309d5a6106b23fe82ea776282d8786ca72ce2be