[roll] Roll fuchsia [starnix] Move linux_root to container
Previously, the `linux_root` mirror of the root file system was exposed
on the kernel component. This design is awkward now that we delay
creating the container until the client actually starts the container.

This CL moves this mirror to be exposed on the container component,
which makes more sense, logically. The CL also renames the mirror to
`fs_root` to avoid confusion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864359
Original-Revision: 9de67cbd3b90d4cdd8c72fd3b4cbe5e03ebc68ac
GitOrigin-RevId: eb0c63318b7bb01247140eb2a9cc00cfb8e35e9a
Change-Id: I700b715f6a1861ac12e672650e15e058aac4c889