[roll] Roll fuchsia [fxblob] Add a new write API to fxfs.fidl
The WriteBlob protocol has a Create method that allows clients to create
a new blob in fxfs with a particular merkle root. If the client sets
the allow_existing flag, the server will overwrite the existing blob.
WriteBlob.Create returns a client end to the BlobWriter protocol. The
client will truncate the blob by calling BlobWriter.GetVmo and will get
a vmo in return. The client will then write blob contents into the
shared vmo and call BlobWriter.BytesReady to signal to the server that
some number of bytes has been written to the vmo.

The reason we separate BlobWriter.GetVmo from WriteBlob.Create is
because pkg-resolver receives the size and contents of the blob upon
making an HTTP get request and we would like to avoid making the HTTP
get request before we're sure that the blob doesn't already exist in
fxblob. We intentionally make the BlobWriter client end and the vmo
unsplittable by deriving the vmo from the BlobWriter protocol instead of
returning both the client end and the vmo from WriteBlob.Create. In this
model, the client end is the single channel that represents the client's
ability to write and is separated from the vmo, which is involved in the
mechanics of writing.

Original-Bug: 125334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849956
Original-Revision: 8bd989ab37c2afe81dbb57323f44b32821bf96cf
GitOrigin-RevId: 6bb13981210040f5577699e028cf2578363118f0
Change-Id: I9fd39db3432e036f85ee1e85965d3563d46457c6