[roll] Roll fuchsia [paver] Use BlockAndController
Add a struct to carry around both the Block and Controller
protocols through the paver codebase.

At the moment the Controller handle is empty, as this involves
updating clients to hand the paver the Controller. This will
be done in a follow up CL.

Original-Bug: 127870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863557
Original-Revision: b702bcd5faaa9f988a45877aa751a0f1391fed9c
GitOrigin-RevId: cbcc4df9ad244cb32da653531f5a1f54defaf16f
Change-Id: I982af37e38e94f88399b3ac0abce7a3716c01a8e