[roll] Roll fuchsia [check-licenses] Add ability to use license classifier
licenseclassifier is an open-source tool that can accurately
detect license texts. It's detection algo is much better than the
strategy that check-licenses currently uses.

This CL adds the licenseclassifier library to golibs,
and hooks it up in check-licenses behind a flag. It will be turned
on in a future CL.

Original-Bug: 106919, 127868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859884
Original-Revision: 3c4510875f11e69023577688d4676d5dcfadbddd
GitOrigin-RevId: f44cc04a27d10f06181503509c3275a932f92417
Change-Id: Ic0ed314552298adb8dcc820e3a2e509a9f3f5571