[roll] Roll fuchsia [netstack3] Add loom infrastructure and test
Add a host-only test suite that uses loom to exhaustively explore the
possible interleaving of threads that are modifying the netstack. This
enables writing tests for atomicity of operations without relying on a
large-scale integration test.

Add a test that asserts the atomicity of packet socket bind operations
as a proof of concept. The test passes, but correctly failed due to a
bug in an earlier version of change
Iadf12d2d9544191cdc680576620951880ab07663.

Original-Bug: 125654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857419
Original-Revision: 16a6c3813f1a078ee252cb8364bef65604a9e34b
GitOrigin-RevId: d27b84a7658df4b832295b7e9eea96a8801bdcf8
Change-Id: I68e51846c40bac0060eb2b285f26a98e0330d5be