[roll] Roll cobalt Remove the custom proto annotations.
The *_1_1 annotations were never used.
The hide_on_client annotation is used on a few fields:
- customer_name and project_name in the MetricDefinition
- expiration_date and owner in the metric Metadata

owner was never used and is removed.
The expiration_date annotation is removed.
customer_name and project_name were convenience fields, and are not
needed and so are removed.

annotations.proto file is not built, but is kept to avoid breaking
internal build rules. It will be removed in fxr/864322.

Original-Bug: b/285217944

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/864318
Original-Revision: d83e0937751bee973388516c9fd903fc31717c42
GitOrigin-RevId: 2663a3b005626b52e5f79cf09f3619e2e813b4ff
Change-Id: I98766869d4764df209b37468546db471129d6ab9