[roll] Roll fuchsia [botanist] Move flash to ffx experiment level 0 and emu to 1.
When prod builders are switched to enabling ffx with experiment level 0,
flashing will be switched on by the migration of prod controllers to
idling in fastboot. For emu, we'll put it back to level 1 until we're
ready to enable in prod.

Original-Bug: 91040, 95938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864941
Original-Revision: 8e8679d2a512c6c9fc1a25052d5bf6c55e2e6690
GitOrigin-RevId: 1a258c688b0f7bb8f4a7ad3fde20b769a3131471
Change-Id: I5d9f12a0b0719b828545db60a78f4a893d72b55e