[roll] Roll fuchsia [build] Use the unversioned ffx binaries for build actions
Build actions should not need the version-stamped ffx
binaries, and using the non-stamped binaries will reduce
the amount of the tree that is dirtied when changing the
git integration commit patch that is stamped into the
binaries.

The result should be fewer targets being needlessly
rebuilt for developers after a jiri update or in CI/CQ
due to the change in the information used to stamp the
binaries.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865021
Original-Revision: 0cb16d6de757e105e624d1fb58eeb717793513ac
GitOrigin-RevId: 2d70f9e376c5ffb5d41c15112e1a6b7ad2d1d5c9
Change-Id: I4ce8efdb4ea42acdc976c7c42966dc531a52fb72