[roll] Roll fuchsia [vfs] Refactor VFS to support node connections in one place
Implementing this for fuchsia.io2 will be simplified if there is one
place that handles node connections. This refactors the Rust VFS library
so that is the case.

This also changes the semantics of NODE_REFERENCE slightly to match how
O_PATH works on Linux: the F_SETFL fcntl should fail with EBADF on a
file opened with O_PATH.

This change also consolidates OpenFile and OpenDirectory into
OpenNode. The two differed previously: OpenFile would support async
close. To reconcile this, we run with the synchronous version of close
which forces Fxfs to spawn a task when it needs to. This actually
uncovered an existing issue where Fxfs would not always flush when
dropping references.

Original-Bug: 77623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862583
Size-Review: David Iserovich <diserovich@google.com>
Original-Revision: aea81d43e0945a41106466f5e53df77de131eee9
GitOrigin-RevId: ecfdc9ee62a38012881201622abc69637e4f9bc4
Change-Id: I3d4e584b3d67019fb5c6830ac1245e2304140929