[roll] Roll fuchsia [display-coordinator] Replace manual memory management with fbl::Vector.
This CL brings a few memory safety improvements to display coordinator's
Controller::DisplayControllerInterfaceOnDisplaysChanged(). Previously,
the code relied on allocating C-style arrays on the heap, and passed the
arrays as unmanaged pointers to an async task closure. This CL migrates
the code to use fbl::Vector.

This CL does not impact the number of heap allocations and the use of
AllocChecker.

Test: fx test display-coordinator-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865025
Original-Revision: a3178881dcea63f423e29e706fc08ca0b0e6c2ed
GitOrigin-RevId: b160efb07183f4e39c6472964a2cdca3329c9a78
Change-Id: I0e3126bfb454f10c52d1071365c401e809d9bcc9