[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Implement 'errno' on the GPU as a global integer internally
The C standard asserts that the `errno` value is an l-value thread local
integer. We cannot provide a generic thread local integer on the GPU
currently without some workarounds. Previously, we worked around this by
implementing the `errno` value as a special consumer class that made all
the writes disappear. However, this is problematic for internal tests.
Currently there are build failures because of this handling and it's
only likely to cause more problems the more we do this.

This patch instead makes the internal target used for testing export the
`errno` value as a simple global integer. This allows us to use and test
the `errno` interface correctly assuming we run with a single thread.
Because this is only used for the non-exported target we still do not
provide this feature in the version that users will use so we do not
need to worrk about it being incorrect in general.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D152015

GitOrigin-RevId: 3a96ae8ed241e0ca9864b525a3be2c215a56c83e
Original-Revision: 3e66a20c5d611459f49e30e0546a02234494179e
Roller-URL: https://ci.chromium.org/b/8779366051070379937
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865618
Original-Revision: fccdc26ddf14af02d816b51ac990eb0d171ceffb
Change-Id: I362bf8f0e1a60464817fcb6b2a0cee1f61c449d5