[roll] Roll chromium/deps/icu@default Add language names for Translate
The following 11 languages are used by Translate so need to be added
to the Accept-Language list:

  ay, bho, bm, doi, dv, ilo, lus, mai, mni, sa, ts

"as" and "kok" will also be added as Translate languages, but are already supported by ICU.

We only need the language name from ICU in all UI languages and the
native name of each of the languages.

The following files are added in source/data/lang to add the native
display name of {ay,dv,ilo,lus,ts} along with the patch file and
a note about them in README.chromium.

ay.txt
dv.txt
ilo.txt
lus.txt
ts.txt

The data size difference is as follows:
         Old      New      Delta
chromeos 11718496 11730784 12288
common   10560960 10573104 12144

Original-Bug: 1313213
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4585451
Original-Revision: 17665248776256842466ab820f5f1ae9637dbfd8
GitOrigin-RevId: 532f21fd4b50b5f26e72efc803e169f33683370b
Change-Id: I72deb7901d240df415424f6daae362f31bbce483