[roll] Roll fuchsia [fuchsia.io] Add a node protocol
Make requesting the node protocol explicit.

Previously, the intention was that an empty table would indicate this
state, but checking for an empty table in FIDL is prone to error, and we
also need to be able to indicate that only a directory is acceptable (in
order to support Posix O_PATH | O_DIRECTORY).

Original-Bug: 77623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863936
Original-Revision: 98933f8456fab52847cddee8df2fc02c2470dcab
GitOrigin-RevId: e26c150017ae2e6fd1fe2affcd4f73527a32d89e
Change-Id: I310f7fc62f225c14cb821dc46bc2e9c96f655ef2