[roll] Roll fuchsia [memory_sampler] Support dynamic memory allocation in hooks
Currently, allocations are forbidden from the scudo
allocation/deallocation hooks, (mostly) because of the risk of
infinite recursion. This commit adds a simple re-entrancy guard, which
is used to detect and prevent such a loop.

Being able to allocate from the hooks is desirable: the usual strategy
to handle deallocations in a sampling profiler is to store, at
allocation time, the sampled allocations in a dynamically allocated
set, and to consult this set at deallocation time to decide which
deallocations need to be reported.

Note that this use case is technically forbidden per Scudo's contract,
though it appears to work well enough, from our manual testing. We
are working with the Scudo owners to lift this restriction. Depending
on where this discussion lands, we might want to revisit this approach
in a subsequent commit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864018
Original-Revision: 4713974f6248be3bda9ea45cabbc2947153db90f
GitOrigin-RevId: 4e32badf4635baa22c07d53567be804e0606f355
Change-Id: Ic79ee8cd1cdca872c13b3a09a75144440c77a274