[roll] Roll fuchsia [fuchsia.paver] Add Controller to UseBlockDevice
Pass the Controller in as well, in preparation for when
fuchsia.hardware.block/Block will stop multiplexing this
protocol.

The controller is not load bearing yet, as there's one
place that still needs to be updated to pass it through.
Most uses of this API can be easily updated to pass the
controller though.

The paver-test needed to be updated to a new
GetNewConnectionsMultiplexed function as there's
still some reliance on multiplexing, but this will
be removed in a followup CL.

Original-Bug: 127870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/863558
Original-Revision: 0d0566cfb38b01c1ecc570abf2ee7a3fcbbac8ca
GitOrigin-RevId: 83b129e71de8b59ffe79fdc2b4983a544928a427
Change-Id: I11489b3d4adb1adf9c98244362d3c594658d2c5f