[roll] Roll fuchsia [starnix][fxfs] re-enable most of select_test
There is one test that somehow manages to make starnix get an invalid
handle and pass it to zx_object_wait_async, causing a starnix kernel
panic. It wasn't immediately obvious how it got in that state so leaving
it as skipped for now to come back to.

Original-Bug: 125830
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867399
Original-Revision: 13e1578f710254496ee90b0abfcb43dd007ba55c
GitOrigin-RevId: 9f16f4fb4834b3e35e16d98887233ee8441fbbe2
Change-Id: Iffb68ccf55e90c0c13ec434db74fe82437b22568