[roll] Roll infra/recipes [breakpad] Statically link rustc_demangle and zlib
The breakpad build has a number of assumptions about the rustc_demangle
that are not universally correct (e.g. the path to librustc_demangle.a)
so we set the appropriate CFLAGS/LDFLAGS in the recipe and skip
`--with-rustc-demangle=...` entirely.

Similarly the breakpad build appends `-lz` at the end of the linker
invocation and we cannot override it with `-l:libz.a` nor does the zlib
build allow us to not build the shared libraries (the install target
depends on them), so simply delete all libz.so* files before building
`dump_syms`.

Verified with `ldd` via led:
https://ci.chromium.org/raw/build/logs.chromium.org/fuchsia/led/nmulcahey_google.com/59bc9a3d38a66d6dd3e2660cea652fbccd707e29c323635462d0ad5a43da93b5/+/build.proto?server=chromium-swarm.appspot.com

Original-Bug: 124645

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/867542
Presubmit-Verified: CQ Bot <fuchsia-internal-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: 0523cd72392b162f8443dedd30ebc08c11782da7
GitOrigin-RevId: 89d1731d6852354915511203ad8b05aade767869
Change-Id: I8895f7baecc1992100bce1f581d6c8ab2f79c258