[roll] Roll fuchsia [remote-control] Add log API
Allow host-side tools to emit a log on the device, similar to the SL4F
logging facade.

Many product E2E tests make use of SL4F the logging facade to do this.
This device-side logging allows test authors to sequence and correlate
specific events in the test with events happening on the device, for
example to know in the device logs where the beginning and the end of a
test case was. That is not easy to do if the log only exist in the test's
output.

It would be a pretty significant regression if Lacewing-based tests
couldn't do the same thing.

Original-Bug: b/276968393 b/276726514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862578
Original-Revision: 73b8d801bfc6cc495694a83814c64b3ba5f3050a
GitOrigin-RevId: f8c0ecec5f736fd7aa02ec618bb09586355c87ba
Change-Id: I8b104e788a7cd94f05fad17b2c59660d930cc7a4