[roll] Roll fuchsia [story_runner] Pass lambda which accepts AnnotationController_WatchAnnotations_Result
UBSan was detecting a function type mismatch on the invoked callback
passed to GetAnnotations. The callback is a fit::function whose argument
expected to be a AnnotationController_GetAnnotations_Result, but the
underlying function that's actually called is a
AnnotationController_WatchAnnotations_Result. Structurally, both result
objects are the same. They're essentially structs that hold a vector of
Annotations.

Original-Bug: 128274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867705
Original-Revision: db8c7369fc52d43c59a84fc07d71f4929f5f6363
GitOrigin-RevId: dcc6c161b23a15d719baa9c9ee8f2bca139d09f0
Change-Id: I20f27f5648077c8a32f57c4f570b9cf9d5f110bd